% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdapterContent.R
\docType{methods}
\name{Adapter_Content,FastqcDataList-method}
\alias{Adapter_Content,FastqcDataList-method}
\alias{Adapter_Content}
\alias{Adapter_Content,FastqcData-method}
\alias{Adapter_Content}
\alias{Adapter_Content,FastqcFile-method}
\alias{Adapter_Content}
\alias{Adapter_Content,FastqcFileList-method}
\alias{Adapter_Content}
\alias{Adapter_Content,character-method}
\alias{Adapter_Content}
\title{Get the Adapter Content information}
\usage{
\S4method{Adapter_Content}{FastqcDataList}(object)

\S4method{Adapter_Content}{FastqcData}(object)

\S4method{Adapter_Content}{FastqcFile}(object)

\S4method{Adapter_Content}{FastqcFileList}(object)

\S4method{Adapter_Content}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files.}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Adapter Content module from one or more FastQC reports
}
\details{
It is currently assumed that all FastQC reports contain information about the same adapters.
If this was varied during the preparation of the reports, this function will probably fail.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Return a data_frame/tibble with the raw information
Adapter_Content(fdl)

}
