% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPercent.R
\name{addPercent}
\alias{addPercent}
\title{Add a percentage sign}
\usage{
addPercent(x)
}
\arguments{
\item{x}{Any vector}
}
\value{
character vector
}
\description{
Add a percentage sign to the end of a string
}
\examples{

x <- 1:10
ngsReports:::addPercent(x)

}
\keyword{internal}
