% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Statistics.R
\docType{methods}
\name{Basic_Statistics,FastqcData-method}
\alias{Basic_Statistics,FastqcData-method}
\alias{Basic_Statistics}
\alias{Basic_Statistics,FastqcDataList-method}
\alias{Basic_Statistics}
\alias{Basic_Statistics,FastqcFile-method}
\alias{Basic_Statistics}
\alias{Basic_Statistics,FastqcFileList-method}
\alias{Basic_Statistics}
\alias{Basic_Statistics,character-method}
\alias{Basic_Statistics}
\title{Get the set of Basic Statistics}
\usage{
\S4method{Basic_Statistics}{FastqcData}(object)

\S4method{Basic_Statistics}{FastqcDataList}(object)

\S4method{Basic_Statistics}{FastqcFile}(object)

\S4method{Basic_Statistics}{FastqcFileList}(object)

\S4method{Basic_Statistics}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Basic Statistics module from one or more FastQC reports
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Return a data_frame/tibble with the raw information
Basic_Statistics(fdl)

}
