% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastpFile.R
\docType{class}
\name{.FastpFile-class}
\alias{.FastpFile-class}
\title{The .FastpFile Object Class}
\arguments{
\item{x}{character(1) denoting a file.path}
}
\value{
An object of class .FastqcFile
}
\description{
The .FastpFile Object Class defines a path to the output from
the standalone tool fastp.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This class simply refers to a fastp output file after checking for
existence and validity (i.e. the correct internal structure).
Underlying files are expected to be in json format

The helper function \code{.FastpFile()} is a simple constructor which
checks validity and enables construction of other dependent classes.
}
\section{Slots}{

\describe{
\item{\code{path}}{Character vector of length 1 which contains a valid file path.}
}}

\keyword{internal}
