% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.getToolName}
\alias{.getToolName}
\title{Identify tool name}
\usage{
.getToolName(x, possTypes)
}
\arguments{
\item{x}{Character vector as output when readLines to a supplied log file}
}
\value{
logical(1)
}
\description{
Identify tool name for log files after
reading in using readLines.
}
\details{
Checks for all the required fields in the lines provided
}
\keyword{internal}
