% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdapterContent.R
\name{.hidePWFRects}
\alias{.hidePWFRects}
\title{Hide PWF tooltips from line plots}
\usage{
.hidePWFRects(x)
}
\arguments{
\item{x}{plotlyObject$x$data}
}
\value{
plotlyObject$x$data
}
\description{
Hide tooltips from PWF rectangles in line plots
}
\keyword{internal}
