% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDupMetrics.R
\name{.isValidDupMetrics}
\alias{.isValidDupMetrics}
\title{Check for a valid Duplication Metrics log}
\usage{
.isValidDupMetrics(x)
}
\arguments{
\item{x}{Character vector containing parsed log file using the function
readLines}
}
\value{
logical(1)
}
\description{
Checks internal structure of the parsed log file
}
\keyword{internal}
