% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.isValidFlagstatLog}
\alias{.isValidFlagstatLog}
\title{Check for correct structure of supplied flagstat}
\usage{
.isValidFlagstatLog(x)
}
\arguments{
\item{x}{Character vector as output when readLines to a supplied file}
}
\value{
logical(1)
}
\description{
Check for correct structure of supplied flagstat files
}
\details{
Checks for all the required fields in the lines provided
}
\keyword{internal}
