% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHisat2Logs.R
\name{.isValidHisat2Log}
\alias{.isValidHisat2Log}
\title{Check for a valid Hisat2 log}
\usage{
.isValidHisat2Log(x)
}
\arguments{
\item{x}{Character vector containing parsed log file using the function
readLines}
}
\value{
logical(1)
}
\description{
Checks internal structure of the parsed log file
}
\keyword{internal}
