% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.isValidMacs2CallpeakLog}
\alias{.isValidMacs2CallpeakLog}
\title{Check for correct structure of macs2 callpeak log}
\usage{
.isValidMacs2CallpeakLog(x)
}
\arguments{
\item{x}{Character vector as output when readLines to a supplied file}
}
\value{
logical(1)
}
\description{
Check for correct structure of macs2 callpeak log
}
\details{
Checks for all the required fields in the lines provided
}
\keyword{internal}
