% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSamtoolsStats.R
\name{.isValidSamtoolsStatsLog}
\alias{.isValidSamtoolsStatsLog}
\title{Check for correct structure of supplied samtools stats files}
\usage{
.isValidSamtoolsStatsLog(x)
}
\arguments{
\item{x}{Character vector as output when readLines to a supplied log file}
}
\value{
logical(1)
}
\description{
Check for correct structure of supplied samtools stats
files after reading in using readLines.
}
\details{
Checks for all the required fields in the lines provided
}
\keyword{internal}
