% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStarLogs.R
\name{.isValidStarLog}
\alias{.isValidStarLog}
\title{Check for a valid Star Alignment log}
\usage{
.isValidStarLog(x)
}
\arguments{
\item{x}{Character vector containing parsed log file using the function
readLines}
}
\value{
logical(1)
}
\description{
Checks internal structure of the parsed log file
}
\keyword{internal}
