% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.isValidUmitoolsDedupLog}
\alias{.isValidUmitoolsDedupLog}
\title{Check for correct structure of umi_tools --dedup log}
\usage{
.isValidUmitoolsDedupLog(x)
}
\arguments{
\item{x}{Character vector as output when readLines to a supplied file}
}
\value{
logical(1)
}
\description{
Check for correct structure of umi_tools --dedup log
}
\details{
Checks for all the required fields in the lines provided
}
\keyword{internal}
