% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.makePwfGradient}
\alias{.makePwfGradient}
\title{Construct a gradient using PwfCols}
\usage{
.makePwfGradient(
  vals,
  pwfCols,
  breaks = c(0, 5, 10, 100),
  passLow = TRUE,
  na.value = "white"
)
}
\arguments{
\item{vals}{The values which need to have the scale generated}

\item{pwfCols}{An object of class PwfCols}

\item{breaks}{The breaks for the PWF bins}

\item{passLow}{Is the PASS category at the low or high end of the numeric
range}

\item{na.value}{The colour to plot for missing values}
}
\value{
Returns a ggplot list
}
\description{
Construct a gradient using PwfCols
}
\details{
This constructs a list of arguments for passing to
\code{scale_fill_gradientn()} using the values contained in the
fill aesthetic and the supplied breaks for PASS/WARN/FAIL criteria.
}
\keyword{internal}
