% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseAdapterRemovalLogs}
\alias{.parseAdapterRemovalLogs}
\title{Parse data from Adapter Removal log files}
\usage{
.parseAdapterRemovalLogs(data, which = 3)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{which}{which element of the log file to return.
Can be 1:4, "sequences", "settings", "statistics" or "distribution"}
}
\value{
tibble
}
\description{
Parse data from Adapter Removal log files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
