% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseBowtieLogs}
\alias{.parseBowtieLogs}
\title{Parse data from Bowtie log files}
\usage{
.parseBowtieLogs(data)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}
}
\value{
data.frame
}
\description{
Parse data from Bowtie log files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
