% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseCutadaptLogs}
\alias{.parseCutadaptLogs}
\title{Parse data from cutadapt log files}
\usage{
.parseCutadaptLogs(data, which = 1)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{which}{which element of the log file to return.
Can be summary, adapter1, adapter2, adapter3 or overview, or any integrer in
1:5}
}
\value{
tibble
}
\description{
Parse data from cutadapt log files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
