% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseDuplicationMetricsLogs}
\alias{.parseDuplicationMetricsLogs}
\title{Parse data from Picard duplicationMetrics log files}
\usage{
.parseDuplicationMetricsLogs(data, which = 1)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{which}{which element of the log file to return.
Can be 1:2, "metrics" or "histogram"}
}
\value{
tibble
}
\description{
Parse data from Picard duplicationMetrics log files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
