% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseFlagstatLogs}
\alias{.parseFlagstatLogs}
\title{Parse data from samtools flagstat files}
\usage{
.parseFlagstatLogs(data, ...)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{...}{Not used}
}
\value{
data.frame
}
\description{
Parse data from samtools flagstat files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
