% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseHisat2Logs}
\alias{.parseHisat2Logs}
\title{Parse data from HISAT2 log files}
\usage{
.parseHisat2Logs(data, ...)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{...}{Not used}
}
\value{
data.frame
}
\description{
Parse data from HISAT2 log files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
