% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseQuastLogs}
\alias{.parseQuastLogs}
\title{Parse data from BUSCO log files}
\usage{
.parseQuastLogs(data, ...)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{...}{Not used}
}
\value{
data.frame
}
\description{
Parse data from BUSCO log files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
