% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseRnaseqcMetricsLogs}
\alias{.parseRnaseqcMetricsLogs}
\title{Parse data from rnaseqc metrics files}
\usage{
.parseRnaseqcMetricsLogs(data, ...)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{...}{Not used}
}
\value{
data.frame
}
\description{
Parse data from rnaseqc metrics files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
