% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseSalmonLibFormatCountsLogs}
\alias{.parseSalmonLibFormatCountsLogs}
\title{Parse data from salmon lib_format_counts.json files}
\usage{
.parseSalmonLibFormatCountsLogs(data, ...)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{...}{Not used}
}
\value{
data.frame
}
\description{
Parse data from salmon lib_format_counts.json files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
