% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseStarLogs}
\alias{.parseStarLogs}
\title{Parse data from STAR log files}
\usage{
.parseStarLogs(data, ...)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{...}{Not used}
}
\value{
tibble
}
\description{
Parse data from STAR log files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
