% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{.parseUmitoolsDedupLogs}
\alias{.parseUmitoolsDedupLogs}
\title{Parse data from umi_tools dedup files}
\usage{
.parseUmitoolsDedupLogs(data, ...)
}
\arguments{
\item{data}{List of lines read using readLines on one or more files}

\item{...}{Not used}
}
\value{
data.frame
}
\description{
Parse data from umi_tools dedup files
}
\details{
Checks for structure will have been performed
}
\keyword{internal}
