% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderDendro.R
\name{.renderDendro}
\alias{.renderDendro}
\title{Set up dendrograms for interactive plots}
\usage{
.renderDendro(df)
}
\arguments{
\item{df}{A `data.frame` as required}
}
\value{
A plotly object
}
\description{
A commonly used (hidden) function for setting up dendrograms
for interactive plots. based on code found at
https://plot.ly/ggplot2/ggdendro-dendrograms/
}
\details{
Create plot using \code{theme_dendro}
}
\keyword{internal}
