% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.updateThemeFromDots}
\alias{.updateThemeFromDots}
\title{Add custom theme elements fro dotArgs}
\usage{
.updateThemeFromDots(p, ...)
}
\arguments{
\item{p}{ggplot object}

\item{...}{Standard dot arguments}
}
\value{
ggplot2 object
}
\description{
Add custom theme elements fro dotArgs
}
\keyword{internal}
