% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyPlot.R
\name{emptyPlot}
\alias{emptyPlot}
\title{Create an empty plot with supplied text}
\usage{
emptyPlot(x)
}
\value{
A ggplot2 object
}
\description{
Create an empty plot with supplied text
}
\details{
Create plot using \code{theme_void} and only with the supplied text
}
\keyword{internal}
