% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportOverrepresented.R
\docType{methods}
\name{exportOverrepresented}
\alias{exportOverrepresented}
\alias{exportOverrepresented,FastqcData-method}
\alias{exportOverrepresented,FastqcData}
\alias{exportOverrepresented,FastqcDataList-method}
\alias{exportOverrepresented,FastqcDataList}
\title{Write fasta of Over-Represented sequences.}
\usage{
exportOverrepresented(x, path, n, labels, noAdapters = TRUE)

\S4method{exportOverrepresented}{FastqcData}(x, path, n = 10, labels,
  noAdapters = TRUE)

\S4method{exportOverrepresented}{FastqcDataList}(x, path, n = 10, labels,
  noAdapters = TRUE)
}
\arguments{
\item{x}{Can be a \code{FastqcData} or \code{FastqcDataList}}

\item{path}{Path to export the fasta file to. Reverts to a default if not supplied}

\item{n}{The number of sequences to output}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{noAdapters}{logical. Remove any sequences identified as possible adapters or primers by FastQC}
}
\value{
Exports to a fasta file, and returns the fasta information invisibly
}
\description{
Output overrepresented sequences to disk in fasta format.
}
\details{
Fasta will contain \code{Filename}, \code{Possible Source}, \code{Percent of total reads}
}
\examples{
\dontrun{
# Get the files included with the package
fileDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(fileDir, pattern = "zip$", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Export the top10 Overrepresented Sequences as a single fasta file
exportOverrepresented(fdl, path = "top10.fa")

}

}
