% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,FastqcFileList,ANY,missing-method}
\alias{[,FastqcFileList,ANY,missing,ANY-method}
\alias{[}
\alias{[,FastqcFileList,character,missing-method}
\alias{[,FastqcFileList,character,missing,ANY-method}
\alias{[}
\alias{[,FastqcDataList,ANY,missing-method}
\alias{[,FastqcDataList,ANY,missing,ANY-method}
\alias{[[}
\alias{[[,FastqcFileList,character,missing-method}
\title{Extract Elements}
\usage{
\S4method{[}{FastqcFileList,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastqcFileList,character,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastqcDataList,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{FastqcFileList,character,missing}(x, i, j, ...)
}
\arguments{
\item{x}{A FastqcFileList or FastqcDataList}

\item{i}{character, logical or integer vector}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}
}
\description{
Extract elements from FastqcFileList Objects
}
\details{
Extract elements in a consistent manner with R conventions
}
