% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastpDataList.R
\docType{class}
\name{FastpDataList-class}
\alias{FastpDataList-class}
\alias{FastpDataList}
\title{The FastpDataList Object Class}
\usage{
FastpDataList(x)
}
\arguments{
\item{x}{Character vector of file paths specifying paths to fastp json output}
}
\value{
An object of class FastpDataList
}
\description{
The FastpDataList Object Class
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\section{Slots}{

\describe{
\item{\code{...}}{this can either be a single character vector of paths to fastp
files, or several instances of .FastpFile objects}
}}

