% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcData.R
\docType{class}
\name{FastqcData-class}
\alias{FastqcData-class}
\title{The FastqcData Object Class}
\value{
An object of class FastqcData
}
\description{
The FastqcData Object Class
}
\section{Slots}{

\describe{
\item{\code{...}}{this can either be a single character vector of paths to FASTQC files, or several instances of FastqcFile objects}
}}

