% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcDataList.R
\docType{class}
\name{FastqcDataList-class}
\alias{FastqcDataList-class}
\title{The FastqcDataList Object Class}
\value{
An object of class FastqcDataList
}
\description{
The FastqcDataList Object Class
}
\section{Slots}{

\describe{
\item{\code{...}}{this can either be a single character vector of paths to FASTQC files, or several instances of FastqcFile objects}
}}

