% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcDataList.R
\docType{methods}
\name{path,FastqcDataList-method}
\alias{path,FastqcDataList-method}
\alias{path,FastqcDataList-methods}
\alias{names,FastqcDataList-method}
\alias{names,FastqcDataList-methods}
\alias{fileNames,FastqcDataList-method}
\alias{fileName,FastqcDataList-method}
\alias{Basic_Statistics,FastqcDataList-method}
\alias{Per_base_sequence_quality,FastqcDataList-method}
\alias{Per_tile_sequence_quality,FastqcDataList-method}
\alias{Per_sequence_quality_scores,FastqcDataList-method}
\alias{Per_base_sequence_content,FastqcDataList-method}
\alias{Per_sequence_GC_content,FastqcDataList-method}
\alias{Per_base_N_content,FastqcDataList-method}
\alias{Sequence_Length_Distribution,FastqcDataList-method}
\alias{Sequence_Duplication_Levels,FastqcDataList-method}
\alias{Overrepresented_sequences,FastqcDataList-method}
\alias{Adapter_Content,FastqcDataList-method}
\alias{Kmer_Content,FastqcDataList-method}
\alias{Total_Deduplicated_Percentage,FastqcDataList-method}
\alias{getSummary,FastqcDataList-method}
\alias{Version,FastqcDataList-method}
\title{Methods for an object of class FastqcDataList}
\usage{
\S4method{path}{FastqcDataList}(object)

\S4method{names}{FastqcDataList}(x)

\S4method{fileNames}{FastqcDataList}(object)

\S4method{Basic_Statistics}{FastqcDataList}(object)

\S4method{Per_base_sequence_quality}{FastqcDataList}(object)

\S4method{Per_tile_sequence_quality}{FastqcDataList}(object)

\S4method{Per_sequence_quality_scores}{FastqcDataList}(object)

\S4method{Per_base_sequence_content}{FastqcDataList}(object)

\S4method{Per_sequence_GC_content}{FastqcDataList}(object)

\S4method{Per_base_N_content}{FastqcDataList}(object)

\S4method{Sequence_Length_Distribution}{FastqcDataList}(object)

\S4method{Sequence_Duplication_Levels}{FastqcDataList}(object)

\S4method{Overrepresented_sequences}{FastqcDataList}(object)

\S4method{Adapter_Content}{FastqcDataList}(object)

\S4method{Kmer_Content}{FastqcDataList}(object)

\S4method{Total_Deduplicated_Percentage}{FastqcDataList}(object)

\S4method{getSummary}{FastqcDataList}(object)

\S4method{Version}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class \code{FastqcDataList}}
}
\value{
Returns a \code{data_frame} for all the methods except \code{path}, \code{names},
\code{fileNames}
}
\description{
Methods and Accessors for an object of class FastqcDataList
}
\details{
Describes all methods for an object of class \code{FastqcDataList}
}
