% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcFile.R
\docType{methods}
\name{FastqcFile,character-method}
\alias{FastqcFile,character-method}
\title{Path to FastQC data from a single fastq file}
\usage{
\S4method{FastqcFile}{character}(filePath)
}
\arguments{
\item{filePath}{The path to a single FastQC output zip file, or uncompressed folder}

\item{object}{An object of class \code{FastqcFile}}
}
\value{
An object of cass \code{FastqcFile}
}
\description{
Define a FastqcFile
}
\details{
Checks the structure of a folder or zip file with the output from FastQC
}
