% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFile-class.R
\docType{class}
\name{FastqcFile-class}
\alias{FastqcFile-class}
\title{The FastqcFile Object Class}
\value{
An object of class FastqcFile
}
\description{
The FastqcFile Object Class
}
\details{
The is an object which refers to a fastqc output file.
Only the path is stored, however the file is checked for the correct structure on formation of the object.
Files can be zipped (*_fastqc.zip) or extracted directories
}
\section{Slots}{

\describe{
\item{\code{path}}{Character vector of length 1 which contains a valid file path.}
}}

