% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcFileList.R
\docType{methods}
\name{FastqcFileList,character-method}
\alias{FastqcFileList,character-method}
\title{Path to FastQC data a vector of fastq files}
\usage{
\S4method{FastqcFileList}{character}(path)
}
\arguments{
\item{path}{The path to a vector of zip files (or uncompressed folders) as output by FastQC}

\item{object}{An object of class \code{FastqcFileList}}
}
\description{
Define a FastqcFileList
}
\details{
Checks the structure of a folder or zip file with the output from FastQC
}
