% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcFileList.R
\docType{methods}
\name{FastqcFileList,character-method}
\alias{FastqcFileList,character-method}
\alias{FastqcFileList,list-method}
\alias{path,FastqcFileList-method}
\alias{fileNames,FastqcFileList-method}
\title{Path to FastQC data a vector of fastq files}
\usage{
\S4method{FastqcFileList}{character}(path)

\S4method{FastqcFileList}{list}(path)

\S4method{path}{FastqcFileList}(object)

\S4method{fileNames}{FastqcFileList}(object)
}
\arguments{
\item{path}{The path to a vector of zip files (or uncompressed folders) as output by FastQC}

\item{object}{An object of class \code{FastqcFileList}}
}
\description{
Define a FastqcFileList

#' @export
#' @rdname FastqcFileList-methods
#' @aliases isCompressed,FastqcFileList-method
setMethod("isCompressed", "FastqcFileList",
          function(object){vapply(object, isCompressed, logical(1))})
}
\details{
Checks the structure of a folder or zip file with the output from FastQC
}
