% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastQC_shinyServer.R
\name{fastqcServer}
\alias{fastqcServer}
\title{Server for fastQC Shiny app}
\usage{
fastqcServer(input, output)
}
\arguments{
\item{input}{Input for shiny app}

\item{output}{Output for shiny UI}
}
\value{
Server data for fastQC shiny.
}
\description{
function to run server side of FastQC shiny
}
\details{
Renders plots, and text
}
