% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastQC_shinyApp.R
\name{fastqcShiny}
\alias{fastqcShiny}
\title{Run fastQC shiny app}
\usage{
fastqcShiny(fastqcInput, subsetAll = "")
}
\arguments{
\item{fastqcInput}{can be a \code{FastqcFileList}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files.}

\item{subsetAll}{a \code{character} vector of length 1 to subset all files input into shiny app by}
}
\value{
UI data for fastQC shiny.
}
\description{
Returns a shiny app interface to parse many fastQC objects
}
\details{
Currently some plots can take a while to render if the \code{FastqcDataList} passed to
\code{fastqcInput} has many elements
}
