% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastQC_shinyUI.R
\docType{data}
\name{fastqcUI}
\alias{fastqcUI}
\title{UI for fastQC Shiny app}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
fastqcUI
}
\value{
UI data for fastQC shiny.
}
\description{
function to run UI side of FastQC shiny
}
\details{
displays plots and text
}
\keyword{datasets}
