% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileName.R
\docType{methods}
\name{fileName,FastqcFile-method}
\alias{fileName,FastqcFile-method}
\title{Return the File Names from an Object}
\usage{
\S4method{fileName}{FastqcFile}(object)
}
\arguments{
\item{object}{An object of class FastqcFile, FastqcFileList, FastqcData or FastqcDataList}
}
\value{
Returns the filenames, without the preceding directories, i.e. basename
}
\description{
The File Names from an object
}
