% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileName.R
\docType{methods}
\name{fileName}
\alias{fileName}
\alias{fileName,FastqcData-method}
\alias{fileName,FastqcDataList-method}
\title{Return the Underlying Fastq File Names from FastqcData* Objects}
\usage{
\S4method{fileName}{FastqcData}(object)

\S4method{fileName}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class FastqcData or FastqcDataList}
}
\value{
Returns the filenames for the Fastq file the FastQC report was
generated from, without any preceding directories.
}
\description{
Return the Underlying Fastq File Names from FastqcData* Objects
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)
fileName(fdl)

}
