% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-StarLogData.R
\docType{methods}
\name{fileNames,StarLogData-method}
\alias{fileNames,StarLogData-method}
\title{Methods for an object of class StarLogData}
\usage{
\S4method{fileNames}{StarLogData}(object)
}
\arguments{
\item{object}{An object of class \code{StarLogData}}
}
\description{
Methods and Accessors for an object of class StarLogData
}
\details{
Describes all methods for an object of class \code{StarLogData}
}
