% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileNames.R
\name{fileNames}
\alias{fileNames}
\title{Return the FileNames from an Object}
\usage{
fileNames(object)
}
\arguments{
\item{object}{An object of class FastqcFile, FastqcFileList, FastqcData or FastqcDataList}
}
\value{
Returns the filenames, without the preceding directories, i.e. basename
}
\description{
The File Names from an object
}
