% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcFromFasta.R
\name{gcFromFasta}
\alias{gcFromFasta}
\title{Calculate GC content from a fasta file}
\usage{
gcFromFasta(Fastafile, n = 1e+06, bp = 100)
}
\arguments{
\item{Fastafile}{A fasta file containing a genome or transcriptome of 
interest}

\item{n}{the number of fragments to simulate}

\item{bp}{the fragment length used in estimation}
}
\value{
A data.frame with the GC percentage and the exected frequency
}
\description{
Calculate GC content from a set of DNA sequences provided by a 
fasta file
}
\details{
Used to obtain theoretical GC content for a given 
genome/transcriptome
}
\author{
Hien To <totuhien@gmail.com>
}
\keyword{internal}
