% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcTheoretical}
\alias{gcTheoretical}
\title{Theoretical GC content}
\format{An object of class \code{TheoreticalGC} of length 1.}
\usage{
gcTheoretical
}
\description{
This object contains the theoretical GC content for each provided species,
for both the genome and transcriptome
}
\details{
The object is defined with the S4 class \code{TheoreticalGC}
}
\keyword{datasets}
