% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalGC.R
\docType{methods}
\name{genomes}
\alias{genomes}
\alias{genomes,TheoreticalGC-method}
\title{List Available Genomes}
\usage{
genomes(object)

\S4method{genomes}{TheoreticalGC}(object)
}
\arguments{
\item{object}{An object of class Theoretical GC}
}
\value{
A \code{data.frame} object
}
\description{
List available genomes in a TheoreticalGC object
}
