% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PwfCols.R
\docType{methods}
\name{getColours,PwfCols-method}
\alias{getColours,PwfCols-method}
\alias{setColours,PwfCols-method}
\alias{setColours}
\alias{setColors}
\alias{setAlpha,PwfCols-method}
\alias{setAlpha}
\title{Get and Set the Colours for a PwfCols Objects}
\usage{
\S4method{getColours}{PwfCols}(x)

\S4method{setColours}{PwfCols}(x, PASS, WARN, FAIL, MAX)

setColours(x, ...)

\S4method{setAlpha}{PwfCols}(x, alpha)
}
\arguments{
\item{x}{An object of class PwfCols}
}
\description{
Define the generics & methods for setting/getting colours from an object of class PwfCols
}
