% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastqcData.R
\docType{methods}
\name{getFastqcData}
\alias{getFastqcData}
\alias{getFastqcData,.FastqcFile-method}
\alias{getFastqcData,NULL-method}
\alias{getFastqcData,character-method}
\title{Get all data from FastQC files}
\usage{
getFastqcData(object)

\S4method{getFastqcData}{.FastqcFile}(object)

\S4method{getFastqcData}{`NULL`}(object)

\S4method{getFastqcData}{character}(object)
}
\arguments{
\item{object}{Can be a .FastqcFile or paths to files}
}
\value{
An object of \code{FastqcData} or a \code{FastqcDataList}
}
\description{
Read the information from the \code{fastqc_data.txt} files in
each .FastqcFile
}
\examples{
# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fileList)

}
