% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastqcData.R
\docType{methods}
\name{getFastqcData,FastqcFile-method}
\alias{getFastqcData,FastqcFile-method}
\title{Get the complete information from FastQC files}
\usage{
\S4method{getFastqcData}{FastqcFile}(object)
}
\arguments{
\item{object}{Can be a FastqcFile or FastqcFileList. FASTQFILELIST IS NOT YET IMPLEMENTED}
}
\value{
An object of \code{FastqcData}
}
\description{
Read the information from the \code{fastqc_data.txt} files in each FastqcFile
}
