% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TheoreticalGC.R
\docType{methods}
\name{getGC,TheoreticalGC-method}
\alias{getGC,TheoreticalGC-method}
\title{Methods for an object of class TheoreticalGC}
\usage{
\S4method{getGC}{TheoreticalGC}(object, species, type)
}
\arguments{
\item{object}{An object of class \code{TheoreticalGC}}

\item{species}{character. The species to extract data for}

\item{type}{character. Can be \code{"Genomic"} or \code{"Transcriptomic"}}
}
\value{
A \code{data_frame}
}
\description{
Methods and Accessors for an object of class TheoreticalGC
}
\details{
Describes all methods for an object of class \code{TheoreticalGC}.
The object \code{\link{gcTheoretical}} supplied with the package contains this
information for a given set of species.

\code{getGC} returns the theoretical GC content for a given species as a data frame.

\code{mData} returns the metadata data frame for all species.

\code{genomes} and \code{transcriptomes} return vectors listing the species for which the
theoretical GC content is available.

\code{allSpecies} returns a vector of the species for which the theoretical GC content is available
}
