% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalGC.R
\docType{methods}
\name{getGC}
\alias{getGC}
\alias{getGC,TheoreticalGC-method}
\title{Get Theoretical GC content}
\usage{
getGC(object, name, type)

\S4method{getGC}{TheoreticalGC}(object, name, type)
}
\arguments{
\item{object}{An object of class Theoretical GC}

\item{name}{The Name of the species in 'Gspecies' format, e.g. Hsapiens}

\item{type}{The type of GC content. Can only be either "Genome" or "Transcriptome"}
}
\value{
A \code{tibble} object
}
\description{
Get the GC content data from a TheoreticalGC object
}
\examples{
getGC(gcTheoretical, name = "Hsapiens", type = "Genome")

}
