% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdend.R
\name{ggdend}
\alias{ggdend}
\title{Set up dendrograms for interactive plots}
\usage{
ggdend(df)
}
\arguments{
\item{df}{A `data.frame` as required}
}
\value{
A plotly object
}
\description{
A commonly used (hidden) function for setting up dendrograms for 
interactive plots
}
\details{
Create plot using \code{theme_dendro}
}
\keyword{internal}
