% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBowtieLogs.R
\name{importBowtieLogs}
\alias{importBowtieLogs}
\title{Import bowtie log files}
\usage{
importBowtieLogs(x)
}
\arguments{
\item{x}{\code{character}. Vector of filenames}
}
\value{
A \code{tibble}.
Column names are broadly similar to the text in supplied files,
but have been modified for easier handling under R naming conventions.
}
\description{
Imports log files generated by stderr output from bowtie
}
\details{
Imports one or more log files as output by the aligner bowtie
}
