% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDupMetrics.R
\name{importDupMetrics}
\alias{importDupMetrics}
\title{Import Metrics From MarkDuplicates}
\usage{
importDupMetrics(x, useBasename = TRUE)
}
\arguments{
\item{x}{\code{character}. Vector of filenames}

\item{useBasename}{\code{logical} Strip file paths from the LIBRARY column
in the returned tibbles?}
}
\value{
A list with two tibbles.
The first (\code{$metrics}) is the metrics for each file, with the second
(\code{$histogram}) being the histogram for each file.
}
\description{
Imports metrics files as generated by MarkDuplicates
}
\details{
Files generated by MarkDuplicates from the picard set of tools can
be imported using this functions.
Files must be exactly as generated by this tool, without modification.
}
\examples{
sysDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(sysDir, "Dedup_metrics.txt", full.names = TRUE)
dupMetrics <- importDupMetrics(fl)

}
