% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHisat2Logs.R
\name{importHisat2Logs}
\alias{importHisat2Logs}
\title{Read Hisat2 Log Files}
\usage{
importHisat2Logs(logs)
}
\arguments{
\item{logs}{\code{character}. Vector of paths to hisat2 log files}
}
\value{
A \code{data_frame}
}
\description{
Import one or more hisat2 log files as a data frame.
}
