% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNgsLogs.R
\name{importNgsLogs}
\alias{importNgsLogs}
\title{Import Various NGS-related log files}
\usage{
importNgsLogs(x, type)
}
\arguments{
\item{x}{\code{character}. Vector of filenames. All log files must be of the
same type. Duplicates of the file name will be ignored.}

\item{type}{\code{character}. The type of file being imported. Can be one of
\code{bowtie}}
}
\value{
A \code{tibble}.
Column names are broadly similar to the text in supplied files,
but have been modified for easier handling under R naming conventions.
}
\description{
Imports NGS-related log files such as those generated from
stderr
}
\details{
Imports one or more log files as output by tools such as:
\code{bowtie}
}
\examples{
f <- c("bowtiePE.txt", "bowtieSE.txt")
bowtieLogs <- system.file("extdata", f, package = "ngsReports")
df <- importNgsLogs(bowtieLogs, type = "bowtie")

}
