% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmer_Content.R
\docType{methods}
\name{Kmer_Content,FastqcData-method}
\alias{Kmer_Content,FastqcData-method}
\alias{Kmer_Content,FastqcDataList-method}
\alias{Kmer_Content,FastqcFile-method}
\alias{Kmer_Content,FastqcFileList-method}
\alias{Kmer_Content,character-method}
\title{Get the Kmer Content information}
\usage{
\S4method{Kmer_Content}{FastqcData}(object)

\S4method{Kmer_Content}{FastqcDataList}(object)

\S4method{Kmer_Content}{FastqcFile}(object)

\S4method{Kmer_Content}{FastqcFileList}(object)

\S4method{Kmer_Content}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Kmer Content module from one or more FastQC reports
}
