% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDendrogram.R
\name{makeDendrogram}
\alias{makeDendrogram}
\title{Set the clusters for heatmap-style plots}
\usage{
makeDendrogram(df, rowVal, colVal, value)
}
\arguments{
\item{df}{The data frame to be clustered}

\item{rowVal}{The rows to be clustered}

\item{colVal}{The value which will become column names}

\item{value}{The value to use for the clustering}
}
\value{
A dendrogram
}
\description{
Set the clusters for heatmap-style plots
}
\examples{
# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)[1:4]
cols <- c("Filename", "Position", "Illumina_Universal_Adapter")
ac <- Adapter_Content(fileList)[cols]
ngsReports:::makeDendrogram(ac, "Filename", "Position", "Illumina_Universal_Adapter")

}
\keyword{internal}
