% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalGC.R
\docType{methods}
\name{mData}
\alias{mData}
\alias{mData,TheoreticalGC-method}
\title{Extract Metadata for TheoreticalGC objects}
\usage{
mData(object)

\S4method{mData}{TheoreticalGC}(object)
}
\arguments{
\item{object}{An object of class Theoretical GC}
}
\value{
A \code{data.frame} object
}
\description{
Extract Metadata for TheoreticalGC objects
}
