% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PwfCols.R
\docType{methods}
\name{names,PwfCols-method}
\alias{names,PwfCols-method}
\title{Define the behaviours for names & names <-}
\usage{
\S4method{names}{PwfCols}(x)
}
\description{
Define the behaviours for names & names <-
}
