% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFile-class.R
\docType{methods}
\name{names<-,FastqcFile,ANY-method}
\alias{names<-,FastqcFile,ANY-method}
\title{Names cannot be set on this type of object. Not required for export}
\usage{
\S4method{names}{FastqcFile,ANY}(x) <- value
}
\description{
Names cannot be set on this type of object. Not required for export
}
