% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overrepresented_sequences.R
\docType{methods}
\name{Overrepresented_sequences,FastqcData-method}
\alias{Overrepresented_sequences,FastqcData-method}
\alias{Overrepresented_sequences}
\alias{Overrepresented_sequences,FastqcDataList-method}
\alias{Overrepresented_sequences,FastqcFile-method}
\alias{Overrepresented_sequences,FastqcFileList-method}
\alias{Overrepresented_sequences,character-method}
\title{Get the Overrepresented Sequences information}
\usage{
\S4method{Overrepresented_sequences}{FastqcData}(object)

\S4method{Overrepresented_sequences}{FastqcDataList}(object)

\S4method{Overrepresented_sequences}{FastqcFile}(object)

\S4method{Overrepresented_sequences}{FastqcFileList}(object)

\S4method{Overrepresented_sequences}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Overrepresented Sequences module from one or more FastQC reports
}
\examples{

# Get the files included with the package
fileDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(fileDir, pattern = "zip$", full.names = TRUE)

# Form a FastqcDataList
fdl <- getFastqcData(fileList)

# Print the Overrepresented Sequences
Overrepresented_sequences(fdl)

}
