% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrepresentedFasta.R
\docType{methods}
\name{overrepresentedFasta}
\alias{overrepresentedFasta}
\alias{overrepresentedFasta,FastqcData-method}
\alias{overrepresentedFasta,FastqcData}
\alias{overrepresentedFasta,FastqcDataList-method}
\alias{overrepresentedFasta,FastqcDataList}
\title{Write fasta of Over-Represented sequences.}
\usage{
overrepresentedFasta(x, outDir = "./", ...)

\S4method{overrepresentedFasta}{FastqcData}(x, outDir = "./", n = 10,
  labels, ...)

\S4method{overrepresentedFasta}{FastqcDataList}(x, outDir = "./", n = 10,
  labels, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcData} or \code{FastqcDataList}}

\item{outDir}{Directory to output fasta files to}

\item{...}{set generic args}

\item{n}{The number of sequences to plot from an individual file}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}
}
\value{
A standard fasta written to \code{outDir}
}
\description{
Output overrepresented sequences to disk in fasta format.
}
\details{
Fasta will contain \code{Filename}, \code{Possible Source}, \code{Percent of total reads}
}
