% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcData.R
\docType{methods}
\name{path,FastqcData-method}
\alias{path,FastqcData-method}
\title{Methods for an object of class FastqcData}
\usage{
\S4method{path}{FastqcData}(object)
}
\arguments{
\item{object}{An object of class \code{FastqcData}}
}
\description{
Methods and Accessors for an object of class FastqcData
}
\details{
Describes all methods for an object of class \code{FastqcData}
}
