% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\docType{methods}
\name{path,FastqcFile-method}
\alias{path,FastqcFile-method}
\title{Return the File Paths from an object}
\usage{
\S4method{path}{FastqcFile}(object)
}
\arguments{
\item{object}{An object of class FastqcFile or FastqcFileList}
}
\value{
A character(1) vector of the file name.
}
\description{
Return the File Paths from an object
}
\details{
Obtains the file.path for objects of multiple classes
}
