% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\docType{methods}
\name{path}
\alias{path}
\alias{path,FastqcFile-method}
\alias{path}
\alias{path,FastqcFileList-method}
\alias{path}
\alias{path,FastqcData-method}
\alias{path}
\alias{path,FastqcDataList-method}
\title{Return the File Paths from an object}
\usage{
\S4method{path}{FastqcFile}(object)

\S4method{path}{FastqcFileList}(object)

\S4method{path}{FastqcData}(object)

\S4method{path}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class FastqcFile or FastqcFileList}
}
\value{
A character vector of the file paths
}
\description{
Return the File Paths from an object
}
\details{
Obtains the file.path for objects of multiple classes
}
