% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Per_base_N_content.R
\docType{methods}
\name{Per_base_N_content,FastqcData-method}
\alias{Per_base_N_content,FastqcData-method}
\alias{Per_base_N_content}
\alias{Per_base_N_content,FastqcDataList-method}
\alias{Per_base_N_content}
\alias{Per_base_N_content,FastqcFile-method}
\alias{Per_base_N_content}
\alias{Per_base_N_content,FastqcFileList-method}
\alias{Per_base_N_content}
\alias{Per_base_N_content,character-method}
\alias{Per_base_N_content}
\title{Get the Per Base N Content information}
\usage{
\S4method{Per_base_N_content}{FastqcData}(object)

\S4method{Per_base_N_content}{FastqcDataList}(object)

\S4method{Per_base_N_content}{FastqcFile}(object)

\S4method{Per_base_N_content}{FastqcFileList}(object)

\S4method{Per_base_N_content}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Per Base N Content module from one or more FastQC reports
}
\examples{

# Get the files included with the package
fileDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(fileDir, pattern = "zip$", full.names = TRUE)

# Form a FastqcDataList
fdl <- getFastqcData(fileList)

# Print any N Content
Per_base_N-content(fdl)

}
