% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Per_base_sequence_quality.R
\docType{methods}
\name{Per_base_sequence_quality,FastqcData-method}
\alias{Per_base_sequence_quality,FastqcData-method}
\alias{Per_base_sequence_quality}
\alias{Per_base_sequence_quality,FastqcDataList-method}
\alias{Per_base_sequence_quality,FastqcFile-method}
\alias{Per_base_sequence_quality,FastqcFileList-method}
\alias{Per_base_sequence_quality,character-method}
\title{Get the Per Base Sequence Quality information}
\usage{
\S4method{Per_base_sequence_quality}{FastqcData}(object)

\S4method{Per_base_sequence_quality}{FastqcDataList}(object)

\S4method{Per_base_sequence_quality}{FastqcFile}(object)

\S4method{Per_base_sequence_quality}{FastqcFileList}(object)

\S4method{Per_base_sequence_quality}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Per Base Sequence Quality module from one or more FastQC reports
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Print the Per_base_sequence_quality
Per_base_sequence_quality(fdl)   

}
