% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Per_sequence_GC_content.R
\docType{methods}
\name{Per_sequence_GC_content,FastqcData-method}
\alias{Per_sequence_GC_content,FastqcData-method}
\alias{Per_sequence_GC_content}
\alias{Per_sequence_GC_content,FastqcDataList-method}
\alias{Per_sequence_GC_content,FastqcFile-method}
\alias{Per_sequence_GC_content,FastqcFileList-method}
\alias{Per_sequence_GC_content,character-method}
\title{Get the Per Sequence GC Content information}
\usage{
\S4method{Per_sequence_GC_content}{FastqcData}(object)

\S4method{Per_sequence_GC_content}{FastqcDataList}(object)

\S4method{Per_sequence_GC_content}{FastqcFile}(object)

\S4method{Per_sequence_GC_content}{FastqcFileList}(object)

\S4method{Per_sequence_GC_content}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{fastqcDataList}, or simply a \code{character} vector
of paths to fastqc files}
}
\value{
A single \code{tibble} containing all information combined from all
supplied FastQC reports
}
\description{
Retrieve the Per Sequence GC Content module from one or more
FastQC reports
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Print the Per_sequence_GC_content
Per_sequence_GC_content(fdl)

}
