% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Per_sequence_quality_scores.R
\docType{methods}
\name{Per_sequence_quality_scores,FastqcData-method}
\alias{Per_sequence_quality_scores,FastqcData-method}
\alias{Per_sequence_quality_scores}
\alias{Per_sequence_quality_scores,FastqcDataList-method}
\alias{Per_sequence_quality_scores}
\alias{Per_sequence_quality_scores,FastqcFile-method}
\alias{Per_sequence_quality_scores}
\alias{Per_sequence_quality_scores,FastqcFileList-method}
\alias{Per_sequence_quality_scores}
\alias{Per_sequence_quality_scores,character-method}
\alias{Per_sequence_quality_scores}
\title{Get the Per Sequence Quality Scores information}
\usage{
\S4method{Per_sequence_quality_scores}{FastqcData}(object)

\S4method{Per_sequence_quality_scores}{FastqcDataList}(object)

\S4method{Per_sequence_quality_scores}{FastqcFile}(object)

\S4method{Per_sequence_quality_scores}{FastqcFileList}(object)

\S4method{Per_sequence_quality_scores}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Per Sequence Quality Scores module from one or more FastQC reports
}
\examples{

# Get the files included with the package
fileDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(fileDir, pattern = "zip$", full.names = TRUE)

# Form a FastqcDataList
fdl <- getFastqcData(fileList)

# Print the Per_sequence_quality_scores
Per_sequence_quality_scores(fdl)   

}
