% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PwfCols.R
\docType{methods}
\name{plot,PwfCols,ANY-method}
\alias{plot,PwfCols,ANY-method}
\title{Methods to simply display colours}
\usage{
\S4method{plot}{PwfCols,ANY}(x, y, ...)
}
\description{
Methods to simply display colours
}
