% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdapterContent.R
\name{plotAdapterContent}
\alias{plotAdapterContent}
\title{Draw an Adapter Content Plot}
\usage{
plotAdapterContent(x, facet = TRUE, nc, ylim, th, trimNames = TRUE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{facet}{logical. Determines whether to use \code{facet_wrap()} or not}

\item{nc}{Number of columns to use when faceting by Adapter Type.
Must be supplied if using facets.}

\item{ylim}{A vector providing limits for the y-axis. Defaults to \code{ylim = c(0, 100)}}

\item{th}{A ggplot2 theme object. Defaults to \code{theme_bw()}}

\item{trimNames}{logical. Remove the file suffix from the legend.}

\item{...}{Not used}
}
\value{
A standard ggplot2 object
}
\description{
Draw an Adapter Content Plot across one or more FASTQC reports
}
\details{
This extracts the Adapter_Content from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.
}
