% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseQualityHeatmap.R
\name{plotBaseQualitiesHeatmap}
\alias{plotBaseQualitiesHeatmap}
\title{Plot the per base quality as a heatmap}
\usage{
plotBaseQualitiesHeatmap(x, subset, type = "Mean", pwfCols,
  dendrogram = FALSE, labels, clusterNames = FALSE, usePlotly = FALSE,
  ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{type}{\code{character} Type of quality data to be presented "mean" or "median"}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for pass, warning, and fail
values in plot}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} is \code{FALSE}
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{clusterNames}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using heirachial clustering}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}
}
\value{
A ggplot2 object
}
\description{
Plot the Per Base Sequence Quality for a set of FASTQC files
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotBaseQualitiesHeatmap(fdl)



}
