% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseQuals.R
\docType{methods}
\name{plotBaseQuals}
\alias{plotBaseQuals}
\alias{plotBaseQuals,character-method}
\alias{plotBaseQuals,character}
\alias{plotBaseQuals,FastqcFile-method}
\alias{plotBaseQuals,FastqcFile}
\alias{plotBaseQuals,FastqcFileList-method}
\alias{plotBaseQuals,FastqcFileList}
\alias{plotBaseQuals,FastqcData-method}
\alias{plotBaseQuals,FastqcData}
\alias{plotBaseQuals,FastqcDataList-method}
\alias{plotBaseQuals,FastqcDataList}
\title{Plot the Base Qualities for each file}
\usage{
plotBaseQuals(x, usePlotly = FALSE, labels, pwfCols, warn = 25,
  fail = 20, boxWidth = 0.8, ...)

\S4method{plotBaseQuals}{character}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 25, fail = 20, boxWidth = 0.8, ...)

\S4method{plotBaseQuals}{FastqcFile}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 25, fail = 20, boxWidth = 0.8, ...)

\S4method{plotBaseQuals}{FastqcFileList}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 25, fail = 20, boxWidth = 0.8, ...)

\S4method{plotBaseQuals}{FastqcData}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 25, fail = 20, boxWidth = 0.8, ...)

\S4method{plotBaseQuals}{FastqcDataList}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 25, fail = 20, boxWidth = 0.8,
  plotType = c("heatmap", "boxplot"), plotValue = c("Mean", "Median"),
  cluster = FALSE, dendrogram = FALSE, nc = 2, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for
pass, warning, and fail values in plot}

\item{warn, fail}{The default values for warn and fail are 30 and 20
respectively (i.e. percentages)}

\item{boxWidth}{set the width of boxes when using a boxplot}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{plotType}{\code{character} Can be either \code{"boxplot"} or
\code{"heatmap"}}

\item{plotValue}{\code{character} Type of data to be presented. Can be
"Mean" or "Median"}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{nc}{\code{numeric}. The number of columns to create in the plot layout.
Only used if drawing boxplots for multiple files in a FastqcDataList}
}
\value{
A standard ggplot2 object
}
\description{
Plot the Base Qualities for each file as separate plots
}
\details{
This replicates the \code{Per base sequence quality} plots from
FASTQC, using facets to plce them all in a single ggplot2 object.

For large datasets, subsetting by R1 or R2 reads may be helpful
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# The default plot for multiple libraries is a heatmap
plotBaseQuals(fdl)

# The default plot for a single library is the standard boxplot
plotBaseQuals(fdl[[1]])

}
