% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombinedBaseQualities.R
\name{plotCombinedBaseQualities}
\alias{plotCombinedBaseQualities}
\title{Plot the combined Per_base_sequence_quality information}
\usage{
plotCombinedBaseQualities(x, subset, value = "Mean", trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{value}{\code{character}. Specify which value whould be plotted.
Can be any of the columns returned by \code{\link{Per_base_sequence_quality}}.
Defaults to \code{value = "Mean"}.
Can additionally set to "all" to plot all available quantities}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot the Per_base_sequence_quality information for a set of FASTQC reports
}
\details{
This enables plotting of any of the supplied values (i.e. Mean, Median etc) across
a set of FASTQC reports.
By default only the Mean will be plotted,
however any number of the supplied values can be added to the plot,
and these will be separated by linetype.
}
