% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDeduplicatedTotals.R
\name{plotDeduplicatedTotals}
\alias{plotDeduplicatedTotals}
\title{Plot the read totals after de-duplication}
\usage{
plotDeduplicatedTotals(x, usePlotly = FALSE, labels, millions,
  bars = "stacked", dupCol = rgb(0.2, 0.2, 0.8), uniqCol = rgb(0.9, 0.2,
  0.2), ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. If \code{TRUE} the plot will render as an interactive plot in the Viewer pane.
Otherwise the plot will be generated using the current plot device.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{millions}{\code{logical}. Use Millions of reads as the scale for the y-axis.
Unless specified, will be set as TRUE automatically if the highest total is > 2e06.}

\item{bars}{\code{character} Can take the values \code{bars = "stacked"} or \code{bars = "adjacent"}.
Specifying \code{bars = "stacked"} will plot stacked bars with Duplicated and Unique reads.
Specifying \code{bars = "adjacent"} will plot Totals before and after de-duplication.}

\item{dupCol}{The colour to use for the Total values}

\item{uniqCol}{The colour to use for the Deduplicated values}

\item{...}{Used to pass additional attributes to theme()}
}
\value{
Returns a ggplot object, or interactive plotly object
}
\description{
Plot the expected read totals after de-duplication
}
\details{
This uses the percentages from T\code{Total_Deduplicated_Percentage} and projects
them using the existing \code{readTotals}.
These numbers are then plotted against the existing readTotals using \code{\link{geom_bar}}

Setting \code{bars = "stacked"} gives unique and duplicated reads as a stacked bar graph
for each file.
Setting \code{bars = "adjacent"} gives the total and deduplicated (i.e. unique) reads as
adjacent bars on the graph for each file.

Plots can be customised using the standard methods of ggplot2
}
\examples{
# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Draw the plot with stacked bars
plotDeduplicatedTotals(fdl)

}
