% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDuplicationLevels.R
\docType{methods}
\name{plotDuplicationLevels}
\alias{plotDuplicationLevels}
\alias{plotDuplicationLevels,character-method}
\alias{plotDuplicationLevels,character}
\alias{plotDuplicationLevels,FastqcFile-method}
\alias{plotDuplicationLevels,FastqcFile}
\alias{plotDuplicationLevels,FastqcFileList-method}
\alias{plotDuplicationLevels,FastqcFileList}
\alias{plotDuplicationLevels,FastqcData-method}
\alias{plotDuplicationLevels,FastqcData}
\alias{plotDuplicationLevels,FastqcDataList-method}
\alias{plotDuplicationLevels,FastqcDataList}
\title{Plot the combined Sequence_Duplication_Levels information}
\usage{
plotDuplicationLevels(x, usePlotly = FALSE, ...)

\S4method{plotDuplicationLevels}{character}(x, usePlotly = FALSE, ...)

\S4method{plotDuplicationLevels}{FastqcFile}(x, usePlotly = FALSE, ...)

\S4method{plotDuplicationLevels}{FastqcFileList}(x, usePlotly = FALSE, ...)

\S4method{plotDuplicationLevels}{FastqcData}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 20, fail = 50, lineCols = c("red", "blue"), ...)

\S4method{plotDuplicationLevels}{FastqcDataList}(x, usePlotly = FALSE,
  deduplication = "pre", clusterNames = FALSE, dendrogram = FALSE, labels,
  pwfCols, heatCol = inferno(50), ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for pass, warning, and fail
values in plot}

\item{warn, fail}{The default values for warn and fail are 20 and 50 respectively (i.e. percentages)}

\item{lineCols}{Colours of the lines drawn for individual libraries}

\item{deduplication}{Plot Duplication levels 'pre' or 'post' deduplication. Can only take values "pre" and "post"}

\item{clusterNames}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} is \code{FALSE}
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{heatCol}{Colour palette used for the heatmap}
}
\value{
A standard ggplot2 or plotly object
}
\description{
Plot the Sequence_Duplication_Levels information for a set of FASTQC reports
}
\details{
This extracts the Sequence_Duplication_Levels from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Draw the default plot for a single file
plotDuplicationLevels(fdl[[1]])

plotDuplicationLevels(fdl)

}
