% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDuplicationLevels.R
\name{plotDuplicationLevels}
\alias{plotDuplicationLevels}
\title{Plot the combined Sequence_Duplication_Levels information}
\usage{
plotDuplicationLevels(x, subset, value = "Mean", trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*", type)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{type}{A regular expression used to filter which value is plotted.
Patterns should match one of \code{\% Total sequences} or \code{\% Deduplicated sequences}}
}
\value{
A standard ggplot2 object
}
\description{
Plot the Sequence_Duplication_Levels information for a set of FASTQC reports
}
\details{
This extracts the Sequence_Duplication_Levels from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.
}
