% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFastqcPCA.R
\docType{methods}
\name{plotFastqcPCA}
\alias{plotFastqcPCA}
\alias{plotFastqcPCA,ANY-method}
\alias{plotFastqcPCA,character-method}
\alias{plotFastqcPCA,FastqcDataList-method}
\title{Draw a PCA plot for Fast QC modules}
\usage{
plotFastqcPCA(x, usePlotly = FALSE, labels, pwfCols, plotValue,
  cluster = FALSE, ...)

\S4method{plotFastqcPCA}{ANY}(x, usePlotly = FALSE, labels, pwfCols,
  plotValue, cluster = FALSE, ...)

\S4method{plotFastqcPCA}{character}(x, usePlotly = FALSE, labels,
  pwfCols, plotValue, cluster = FALSE, ...)

\S4method{plotFastqcPCA}{FastqcDataList}(x, usePlotly = FALSE, labels,
  pwfCols, plotValue, cluster = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{usePlotly}{\code{logical}. Output as ggplot2 (default) or plotly
object.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours
for PASS/WARN/FAIL}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{modules}{\code{character} vector containing 
the desired FastQC module (eg. c("Per_base_sequence_quality","Per_base_sequence_content"))}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Draw a PCA plot for Fast QC modules across multiple samples
}
\details{
This carries out PCA on all or a subset of FastQC modules and plots the
output using either ggplot or plotly. Clustering of the PCA can be carried
out using a k-means approach.
}
\examples{


}
