% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGcContent.R
\docType{methods}
\name{plotGcContent}
\alias{plotGcContent}
\alias{plotGcContent,ANY-method}
\alias{plotGcContent,character-method}
\alias{plotGcContent,FastqcData-method}
\alias{plotGcContent,FastqcDataList-method}
\title{Plot the Per Sequence GC Content}
\usage{
plotGcContent(
  x,
  usePlotly = FALSE,
  labels,
  theoreticalGC = TRUE,
  gcType = c("Genome", "Transcriptome"),
  species = "Hsapiens",
  GCobject,
  Fastafile,
  n = 1e+06,
  ...
)

\S4method{plotGcContent}{ANY}(
  x,
  usePlotly = FALSE,
  labels,
  theoreticalGC = TRUE,
  gcType = c("Genome", "Transcriptome"),
  species = "Hsapiens",
  GCobject,
  Fastafile,
  n = 1e+06,
  ...
)

\S4method{plotGcContent}{character}(
  x,
  usePlotly = FALSE,
  labels,
  theoreticalGC = TRUE,
  gcType = c("Genome", "Transcriptome"),
  species = "Hsapiens",
  GCobject,
  Fastafile,
  n = 1e+06,
  ...
)

\S4method{plotGcContent}{FastqcData}(
  x,
  usePlotly = FALSE,
  labels,
  theoreticalGC = TRUE,
  gcType = c("Genome", "Transcriptome"),
  species = "Hsapiens",
  GCobject,
  Fastafile,
  n = 1e+06,
  counts = FALSE,
  lineCols = c("red", "blue"),
  ...
)

\S4method{plotGcContent}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  theoreticalGC = TRUE,
  gcType = c("Genome", "Transcriptome"),
  species = "Hsapiens",
  GCobject,
  Fastafile,
  n = 1e+06,
  plotType = c("heatmap", "line", "cdf"),
  pwfCols,
  cluster = FALSE,
  dendrogram = FALSE,
  ...
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or character
vector of file paths}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{theoreticalGC}{\code{logical} default is \code{FALSE} to give the true
GC content, set to \code{TRUE} to normalize values of GC_Content by the
theoretical values using \code{\link[=gcTheoretical]{gcTheoretical()}}. \code{species} must be
specified.}

\item{gcType}{\code{character} Select type of data to normalize GC
content against. Accepts either "Genome" (default) or "Transcriptome".}

\item{species}{\code{character} if \code{gcTheory} is \code{TRUE} it must be
accompanied by a species. Species currently supported can be obtained using
\code{mData(gcTheoretical)}}

\item{GCobject}{an object of class GCTheoretical.
Defaults to the gcTheoretical object supplied with the package}

\item{Fastafile}{a fasta file contains DNA sequences to generate theoretical
GC content}

\item{n}{number of simulated reads to generate theoretical GC content from
\code{Fastafile}}

\item{...}{Used to pass various potting parameters to theme.}

\item{counts}{\code{logical}. Plot the counts from each file if
\code{counts = TRUE}, otherwise frequencies will be plotted.
Ignored if calling the function on a FastqcDataList.}

\item{lineCols}{Colors for observed and theoretical GC lines in single plots}

\item{plotType}{Takes values "line", "heatmap" or "cdf"}

\item{pwfCols}{Object of class \code{\link[=PwfCols]{PwfCols()}} to give colours for
pass, warning, and fail values in plot}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram  will be displayed.}
}
\value{
A ggplot2 or plotly object
}
\description{
Plot the Per Sequence GC Content for a set of FASTQC files
}
\details{
Makes plots for GC_Content.
When applied to a single FastqcData object a simple line plot will be drawn,
with Theoretical GC content overlaid if desired.

When applied to multiple FastQC reports, the density at each GC content bin
can be shown as a heatmap by setting \code{theoreticalGC = FALSE}. By
default the difference in observed and expected theoretical GC is shown.
Species and genome/transcriptome should also be set if utilising the
theoretical GC content.

As an alternative to a heatmap, a series of overlaid distributions can be
shown by setting \code{plotType = "line"}.

Can produce a static ggplot2 object or an interactive plotly object.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# The default plot for a FastqcDataList
plotGcContent(fdl)

# Plot a single FastqcData object
plotGcContent(fdl[[1]])

}
