% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGcContent.R
\name{plotGcContent}
\alias{plotGcContent}
\title{Plot the Per Sequence GC Content}
\usage{
plotGcContent(x, usePlotly = FALSE, labels, counts = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{counts}{\code{logical}. Plot the counts from each file if \code{counts = TRUE}.
If \code{counts = FALSE} the frequencies will be plotted}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}
}
\value{
A plotly object

A ggplot2 object
}
\description{
Plot the Per Sequence GC Content for a set of FASTQC files
}
\details{
Draws a plot of all GC Content overlaid across a single plot.
The mean from the observed data at each value is overlaid as a dotted black line.
Addition of a theoretical distribution has not yet been implemented.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotGcContent(fdl)

# Plot the R1 files using counts
r1 <- grepl("R1", fileName(fdl))


}
