% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGcHeatmap.R
\name{plotGcHeatmap}
\alias{plotGcHeatmap}
\title{Plot GC Content for each file}
\usage{
plotGcHeatmap(x, usePlotly = FALSE, counts = FALSE, clusterNames = FALSE,
  pwfCols, labels, GCtheory = FALSE, GCtheoryType = "Genome",
  species = "Hsapiens", GCobject, dendrogram = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{counts}{\code{logical}. Display counts of GC content rather than frequency}

\item{clusterNames}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using heirachial clustering}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for pass, warning, and fail
values in plot}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{GCtheory}{\code{logical} default is \code{FALSE} to give the true GC content%, set to \code{TRUE} to normalize
values of GC_Content by the theoretical values using \code{\link{gcTheoretical}}. \code{species} must be specified.}

\item{GCtheoryType}{\code{"character"} Select type of data to normalize GC content% agianst accepts either "Genome" or
"Transcriptome" Default is "Genome"}

\item{species}{\code{character} if \code{gcTheory} is \code{TRUE} its must be accompanied by a species
Species currently supported can be obtained using \code{mData(gcTheoretical)}}

\item{GCobject}{an object of class GCTheoretical.
Defaults to the gcTheoretical object supplied witht= the package}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} is \code{FALSE}
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}
}
\value{
A plotly object
}
\description{
Plot the GC content for each file
}
\details{
makes interactive heatmap for \code{GC content}

For large datasets, subsetting by R1 or R2 reads may be helpful
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotGcHeatmap(fdl)

# Using counts
plotGcHeatmap(fdl, counts = TRUE)


}
