% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKmers.R
\name{plotKmers}
\alias{plotKmers}
\title{Plot Overrepresented Kmers}
\usage{
plotKmers(x, subset, nc = 2, nKmers = 6, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{nc}{\code{numeric}. The number of columns to create in the plot layout}

\item{nKmers}{\code{numeric}.
The number of Kmers to show.
Kmers are sorted by \code{max(Obs/Exp_Max)} acros all files to select these}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot Overrepresented Kmers across an entire library
}
