% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNContent.R
\name{plotNContent}
\alias{plotNContent}
\title{Plot the Per Base N Content}
\usage{
plotNContent(x, subset, pass = rgb(0, 0.8, 0), warn = rgb(0.9, 0.9, 0.2),
  fail = rgb(0.8, 0.2, 0.2), trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{pass}{The colour to use for low percentages}

\item{warn}{The colour to use at the WARN threshold}

\item{fail}{The colour to use at the FAIL threshold}

\item{trimNames}{\code{logical}. Remove the file suffix from the names displyed in the legend.}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot the Per Base N Content for a set of FASTQC files as a heatmap
}
\details{
Draws a heatmap based on the N content at every position.
Three colours are used (green, yellow, red) set at the values 0, 5 and 20 respectively.
According to the FASTQC help page values above 5 will issue a WARN,
whilst those above 20 will issue a FAIL
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "fastqcReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Draw the default plot
plotNContent(fdl)



}
