% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNContentPlotly.R
\name{plotNContentPlotly}
\alias{plotNContentPlotly}
\title{Plot the Per base N content as a Heatmap}
\usage{
plotNContentPlotly(x, subset, pwfCols, pattern = "(.+)\\\\.(fastq|fq).*",
  clusterNames = FALSE, trimNames = TRUE, dendrogram = FALSE,
  usePlotly = FALSE)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{clusterNames}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using heirachial clustering}

\item{trimNames}{\code{logical}. Remove the file suffix from the names displyed in the legend.}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} and \code{usePlotly} are \code{FALSE}.
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{pwfcols}{Object of class \code{\link{Pwfcol}} to give colours for pass, warning, and fail
values in plot}
}
\value{
A ggplot2 or plotly object
}
\description{
Plot the Per Base N Content for a set of FASTQC files
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotGcHeatmap(fdl)

# Using counts
plotGcHeatmap(fdl, counts = TRUE)



}
