% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverrep.R
\docType{methods}
\name{plotOverrep}
\alias{plotOverrep}
\alias{plotOverrep,character-method}
\alias{plotOverrep,character}
\alias{plotOverrep,FastqcFile-method}
\alias{plotOverrep,FastqcFile}
\alias{plotOverrep,FastqcFileList-method}
\alias{plotOverrep,FastqcFileList}
\alias{plotOverrep,FastqcData-method}
\alias{plotOverrep,FastqcData}
\alias{plotOverrep,FastqcDataList-method}
\alias{plotOverrep,FastqcDataList}
\title{Plot a summary of Over-represented Sequences}
\usage{
plotOverrep(x, usePlotly = FALSE, labels, pwfCols, ...)

\S4method{plotOverrep}{character}(x, usePlotly = FALSE, labels, pwfCols,
  ...)

\S4method{plotOverrep}{FastqcFile}(x, usePlotly = FALSE, labels, pwfCols,
  ...)

\S4method{plotOverrep}{FastqcFileList}(x, usePlotly = FALSE, labels,
  pwfCols, ...)

\S4method{plotOverrep}{FastqcData}(x, usePlotly = FALSE, labels, pwfCols,
  n = 10, ...)

\S4method{plotOverrep}{FastqcDataList}(x, usePlotly = FALSE, labels,
  pwfCols, cluster = TRUE, dendrogram = TRUE, ...,
  paletteName = "Set1", expand.x = c(0, 0), expand.y = c(0, 0))
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{FastqcDataList} or file path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours
for PASS/WARN/FAIL}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{n}{The number of sequences to plot from an individual file}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{paletteName}{Name of the palette for colouring the possible sources
of the overrepresented sequences. Must be a palette name from
\code{RColorBrewer}}

\item{expand.x, expand.y}{Vectors of length 2. Passed to
\code{scale_*_continuous()}}
}
\value{
A standard ggplot2 object
}
\description{
Plot a summary of Over-represented Sequences for a set of
FASTQC reports
}
\details{
Percentages are obtained by simply summing those within a report.
Any possible double counting by FastQC is ignored for the purposes of a
simple approximation.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Another example which isn't ideal
plotOverrep(fdl)

}
