% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverrepresentedHeatmap.R
\name{plotOverrepresentedHeatmap}
\alias{plotOverrepresentedHeatmap}
\title{Plot Overrepresented Sequnces As a Heatmap}
\usage{
plotOverrepresentedHeatmap(x, subset, nSeq = 20, low = rgb(0.2, 0, 0.2),
  high = rgb(1, 0, 0), naCol = "grey80", flip = TRUE, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{nSeq}{\code{numeric}.
The number of Sequences to show.
Sequences are sorted by \code{max(Percentage)} across all files during selection}

\item{low}{colour used as the low colour in the heatmap}

\item{high}{colour used as the high colour in the heatmap}

\item{naCol}{colour used for missing values}

\item{flip}{\code{logical}. Enable a call to \code{coord_flip} to determine the best direction}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot Overrepresented Sequences across an entire library
}
