% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverrepresentedSummary.R
\docType{methods}
\name{plotOverrepresentedSummary}
\alias{plotOverrepresentedSummary}
\alias{plotOverrepresentedSummary,character-method}
\alias{plotOverrepresentedSummary,character}
\alias{plotOverrepresentedSummary,FastqcFile-method}
\alias{plotOverrepresentedSummary,FastqcFile}
\alias{plotOverrepresentedSummary,FastqcFileList-method}
\alias{plotOverrepresentedSummary,FastqcFileList}
\alias{plotOverrepresentedSummary,FastqcData-method}
\alias{plotOverrepresentedSummary,FastqcData}
\alias{plotOverrepresentedSummary,FastqcDataList-method}
\alias{plotOverrepresentedSummary,FastqcDataList}
\title{Plot a summary of Over-represented Sequences}
\usage{
plotOverrepresentedSummary(x, usePlotly = FALSE, ...)

\S4method{plotOverrepresentedSummary}{character}(x, usePlotly = FALSE, ...)

\S4method{plotOverrepresentedSummary}{FastqcFile}(x, usePlotly = FALSE, ...)

\S4method{plotOverrepresentedSummary}{FastqcFileList}(x, usePlotly = FALSE,
  ...)

\S4method{plotOverrepresentedSummary}{FastqcData}(x, usePlotly = FALSE,
  labels, col = rgb(0.4, 0.4, 0.4, 0.4), ...)

\S4method{plotOverrepresentedSummary}{FastqcDataList}(x, usePlotly = FALSE,
  labels, sequenceSource = "(Primer|Adapter)", col1 = rgb(0.2, 0.2, 0.8),
  col2 = rgb(0.9, 0.2, 0.2), ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{...}{other args}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{col}{The colour to use for the fastqcData version of the plot.}

\item{sequenceSource}{A regular expression used to group the possible overrepresented sequence source into those
matching the pattern, and "Other".
Those matching will be classified using the pattern itself, without any supplied braces.
Defaults to \code{sequenceSource = "(Primer|Adapter)"}}

\item{col1}{The colour to use for the 'Other' percentage}

\item{col2}{The colour to use for the percentage matching \code{sequenceSource}}
}
\value{
A standard ggplot2 object
}
\description{
Plot a summary of Over-represented Sequences for a set of FASTQC reports
}
\details{
Percentages are obtained by simply summing those within a report.
Any possible double counting by FastQC is ignored for the purposes of a simple approximation.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Another example which isn't ideal
plotOverrepresentedSummary(fdl)

}
