% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverrepresentedSummary.R
\name{plotOverrepresentedSummary}
\alias{plotOverrepresentedSummary}
\title{Plot a summary of OVer-represented Sequences}
\usage{
plotOverrepresentedSummary(x, subset, source = "(Primer|Adapter)",
  col1 = rgb(0.2, 0.2, 0.8), col2 = rgb(0.9, 0.2, 0.2), trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{source}{A regular expression used to group the possible source into those
matching the pattern, and "Other".
Those matching will be classified using the pattern itself, without any supplied braces.
Defaults to \code{source = "(Primer|Adapter)"}}

\item{col1}{The colour to use for the Total values}

\item{col2}{The colour to use for the Deduplicated values}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot a summary of OVer-represented Sequences for a set of FASTQC reports
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Another example which isn't ideal
plotOverrepresentedSummary(fdl)

}
