% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeqContent.R
\docType{methods}
\name{plotSeqContent}
\alias{plotSeqContent}
\alias{plotSeqContent,character-method}
\alias{plotSeqContent,character}
\alias{plotSeqContent,FastqcFile-method}
\alias{plotSeqContent,FastqcFile}
\alias{plotSeqContent,FastqcFileList-method}
\alias{plotSeqContent,FastqcFileList}
\alias{plotSeqContent,FastqcData-method}
\alias{plotSeqContent,FastqcData}
\alias{plotSeqContent,FastqcDataList-method}
\alias{plotSeqContent,FastqcDataList}
\title{Plot the per base content as a heatmap}
\usage{
plotSeqContent(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{character}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{FastqcFile}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{FastqcFileList}(x, usePlotly = FALSE, labels,
  ...)

\S4method{plotSeqContent}{FastqcData}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{FastqcDataList}(x, usePlotly = FALSE, labels,
  pwfCols, plotType = c("heatmap", "line"), cluster = TRUE,
  dendrogram = TRUE, ..., nc = 2)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{FastqcDataList} or file path}

\item{usePlotly}{\code{logical}. Generate an interactive plot using plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for
pass, warning, and fail
values in plot}

\item{plotType}{\code{character}. Type of plot to generate. Must be "line" or
"heatmap"}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{nc}{Specify the number of columns if plotting a FastqcDataList as line
plots}
}
\value{
A ggplot2 object
}
\description{
Plot the Per Base content for a set of FASTQC files.
Informative plot where per base sequence content (%A, %T, %G, %C),
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# The default plot
plotSeqContent(fdl)


}
