% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeqQuals.R
\docType{methods}
\name{plotSeqQuals}
\alias{plotSeqQuals}
\alias{plotSeqQuals,character-method}
\alias{plotSeqQuals,character}
\alias{plotSeqQuals,FastqcFile-method}
\alias{plotSeqQuals,FastqcFile}
\alias{plotSeqQuals,FastqcFileList-method}
\alias{plotSeqQuals,FastqcFileList}
\alias{plotSeqQuals,FastqcData-method}
\alias{plotSeqQuals,FastqcData}
\alias{plotSeqQuals,FastqcDataList-method}
\alias{plotSeqQuals,FastqcDataList}
\title{Plot the Per Sequence Quality Scores}
\usage{
plotSeqQuals(x, usePlotly = FALSE, labels, pwfCols, counts = FALSE,
  alpha = 0.1, warn = 30, fail = 20, ...)

\S4method{plotSeqQuals}{character}(x, usePlotly = FALSE, labels, pwfCols,
  counts = FALSE, alpha = 0.1, warn = 30, fail = 20, ...)

\S4method{plotSeqQuals}{FastqcFile}(x, usePlotly = FALSE, labels,
  pwfCols, counts = FALSE, alpha = 0.1, warn = 30, fail = 20, ...)

\S4method{plotSeqQuals}{FastqcFileList}(x, usePlotly = FALSE, labels,
  pwfCols, counts = FALSE, alpha = 0.1, warn = 30, fail = 20, ...)

\S4method{plotSeqQuals}{FastqcData}(x, usePlotly = FALSE, labels,
  pwfCols, counts = FALSE, alpha = 0.1, warn = 30, fail = 20, ...)

\S4method{plotSeqQuals}{FastqcDataList}(x, usePlotly = FALSE, labels,
  pwfCols, counts = FALSE, alpha = 0.1, warn = 30, fail = 20,
  plotType = c("heatmap", "line"), dendrogram = FALSE,
  cluster = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours
for PASS/WARN/FAIL}

\item{counts}{\code{logical}. Plot the counts from each file if
\code{counts = TRUE}, otherwise the frequencies will be plotted}

\item{alpha}{set alpha for line graph bounds}

\item{warn, fail}{The default values for warn and fail are 5 and 10
respectively (i.e. percentages)}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}

\item{plotType}{\code{character}. Can only take the values
\code{plotType = "heatmap"} or \code{plotType = "line"}}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Plot the Per Sequence Quality Scores for a set of FASTQC reports
}
\details{
Plots the distribution of average sequence quality scores across the
set of files. Values can be plotted either as counts (\code{counts = TRUE})
or as frequencies (\code{counts = FALSE}).

Any faceting or scale adjustment can be performed after generation of the
initial plot, using the standard methods of ggplot2 as desired.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# The default plot
plotSeqQuals(fdl)

# Also subset the reads to just the R1 files
r1 <- grepl("R1", fileName(fdl))
plotSeqQuals(fdl[r1])


}
