% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequenceContent.R
\docType{methods}
\name{plotSequenceContent}
\alias{plotSequenceContent}
\alias{plotSequenceContent,character-method}
\alias{plotSequenceContent,character}
\alias{plotSequenceContent,FastqcFile-method}
\alias{plotSequenceContent,FastqcFile}
\alias{plotSequenceContent,FastqcFileList-method}
\alias{plotSequenceContent,FastqcFileList}
\alias{plotSequenceContent,FastqcData-method}
\alias{plotSequenceContent,FastqcData}
\alias{plotSequenceContent,FastqcDataList-method}
\alias{plotSequenceContent,FastqcDataList}
\title{Plot the per base content as a heatmap}
\usage{
plotSequenceContent(x, usePlotly = FALSE, labels, ...)

\S4method{plotSequenceContent}{character}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSequenceContent}{FastqcFile}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSequenceContent}{FastqcFileList}(x, usePlotly = FALSE, labels,
  ...)

\S4method{plotSequenceContent}{FastqcData}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSequenceContent}{FastqcDataList}(x, usePlotly = FALSE, labels,
  plotType = c("heatmap", "line"), pwfCols, cluster = FALSE,
  dendrogram = TRUE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. Generate an interactive plot using plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{plotType}{\code{character}. Type of plot to generate. Must be "line" or "heatmap"}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for pass, warning, and fail
values in plot}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}
}
\value{
A ggplot2 object
}
\description{
Plot the Per Base content for a set of FASTQC files.
Informative plot where per base sequence content (%A, %T, %G, %C),
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotSequenceContent(fdl)


}
