% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequenceContentHeatmap.R
\name{plotSequenceContent}
\alias{plotSequenceContent}
\title{Plot the per base content as a heatmap}
\usage{
plotSequenceContent(x, subset)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}
}
\value{
A ggplot2 object
}
\description{
Plot the Per Base content for a set of FASTQC files.
Informative plot where per base sequence content (%A, %T, %G, %C),
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotGcHeatmap(fdl)

# Using counts
plotGcHeatmap(fdl, counts = TRUE)



}
