% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequenceLengthDistribution.R
\docType{methods}
\name{plotSequenceLengthDistribution}
\alias{plotSequenceLengthDistribution}
\alias{plotSequenceLengthDistribution,character-method}
\alias{plotSequenceLengthDistribution,character}
\alias{plotSequenceLengthDistribution,FastqcFile-method}
\alias{plotSequenceLengthDistribution,FastqcFile}
\alias{plotSequenceLengthDistribution,FastqcFileList-method}
\alias{plotSequenceLengthDistribution,FastqcFileList}
\alias{plotSequenceLengthDistribution,FastqcData-method}
\alias{plotSequenceLengthDistribution,FastqcData}
\alias{plotSequenceLengthDistribution,FastqcDataList-method}
\alias{plotSequenceLengthDistribution,FastqcDataList}
\title{Plot the Sequence Length Distribution}
\usage{
plotSequenceLengthDistribution(x, usePlotly = FALSE, ...)

\S4method{plotSequenceLengthDistribution}{character}(x, usePlotly = FALSE,
  ...)

\S4method{plotSequenceLengthDistribution}{FastqcFile}(x, usePlotly = FALSE,
  ...)

\S4method{plotSequenceLengthDistribution}{FastqcFileList}(x,
  usePlotly = FALSE, ...)

\S4method{plotSequenceLengthDistribution}{FastqcData}(x, usePlotly = FALSE,
  labels, ..., expand.x = c(0, 0.2))

\S4method{plotSequenceLengthDistribution}{FastqcDataList}(x,
  usePlotly = FALSE, labels, counts = FALSE, plotType = "heatmap",
  cluster = FALSE, dendrogram = FALSE, ..., expand.x = c(0, 0.2),
  gridlineCol = "grey20", gridlineWidth = 0.2, heatCol = inferno(50))
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. Output as ggplot2 or plotly object.}

\item{...}{Used to pass additional attributes to theme()}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{expand.x}{Passed to \code{scale_x_discrete}}

\item{counts}{\code{logical} Should distributions be shown as counts or frequencies (percentages)}

\item{plotType}{\code{character}. Can only take the values \code{plotType = "heatmap"}
or \code{plotType = "line"}}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} and \code{usePlotly} are \code{FALSE}.
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{gridlineWidth, gridlineCol}{Passed to geom_hline and geom_vline to determine
width and colour of gridlines}

\item{heatCol}{The colour scheme for the heatmap}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Plot the Sequence Length Distribution across one or more FASTQC reports
}
\details{
This extracts the Sequence Length Distribution from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.
For example, preset axis limits can also be overwritten easily by adding a call to \code{scale_y_continuous}
after the call to \code{plotSequenceLengthDistribution}.

An alternative interactive plot is available by setting the argument \code{usePlotly = TRUE}.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Plot as a frequency plot using lines
plotSequenceLengthDistribution(fdl)

}
