% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequenceLengthDistribution.R
\name{plotSequenceLengthDistribution}
\alias{plotSequenceLengthDistribution}
\title{Draw an Adapter Content Plot}
\usage{
plotSequenceLengthDistribution(x, usePlotly = FALSE, labels, counts = FALSE,
  plotType = "heatmap", ..., expand.x = c(0, 0.2))
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. Output as ggplot2 or plotly object.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{counts}{\code{logical} Should distributions be shown as counts or frequencies (percentages)}

\item{plotType}{\code{character}. Can only take the values \code{plotType = "heatmap"}
or \code{plotType = "line"}}

\item{...}{Used to pass additional attributes to theme()}

\item{expand.x}{Passed to \code{scale_x_discrete}}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Draw an Adapter Content Plot across one or more FASTQC reports
}
\details{
This extracts the Adapter_Content from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.


Preset axis limits can also be overwritten easily by adding a call to \code{scale_y_continuous}
after the call to \code{plotAdapterContent}.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Plot as a frequency plot using lines
plotSequenceLengthDistribution(fdl)

}
