% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequenceQualities.R
\docType{methods}
\name{plotSequenceQualities}
\alias{plotSequenceQualities}
\alias{plotSequenceQualities,character-method}
\alias{plotSequenceQualities,character}
\alias{plotSequenceQualities,FastqcFile-method}
\alias{plotSequenceQualities,FastqcFile}
\alias{plotSequenceQualities,FastqcFileList-method}
\alias{plotSequenceQualities,FastqcFileList}
\alias{plotSequenceQualities,FastqcData-method}
\alias{plotSequenceQualities,FastqcData}
\alias{plotSequenceQualities,FastqcDataList-method}
\alias{plotSequenceQualities,FastqcDataList}
\title{Plot the Per Sequence Quality Scores}
\usage{
plotSequenceQualities(x, usePlotly = FALSE, ...)

\S4method{plotSequenceQualities}{character}(x, usePlotly = FALSE, ...)

\S4method{plotSequenceQualities}{FastqcFile}(x, usePlotly = FALSE, ...)

\S4method{plotSequenceQualities}{FastqcFileList}(x, usePlotly = FALSE, ...)

\S4method{plotSequenceQualities}{FastqcData}(x, usePlotly = FALSE,
  counts = FALSE, warn = 30, fail = 20, pwfCols, alpha = 0.2, ...)

\S4method{plotSequenceQualities}{FastqcDataList}(x, usePlotly = FALSE,
  counts = FALSE, pwfCols, labels, plotType = "heatmap",
  dendrogram = FALSE, clusterNames = FALSE, lineCol = "grey20",
  lineWidth = 0.2, alpha = 0.1, warn = 30, fail = 20, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}

\item{counts}{\code{logical}. Plot the counts from each file if \code{counts = TRUE}.
If \code{counts = FALSE} the frequencies will be plotted}

\item{warn, fail}{The default values for warn and fail are 5 and 10 respectively (i.e. precentages)}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{alpha}{set alpha for line graph bounds}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{plotType}{\code{character}. Can only take the values \code{plotType = "heatmap"}
or \code{plotType = "line"}}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} is \code{FALSE}
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{clusterNames}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{lineWidth, lineCol}{Passed to geom_hline and geom_vline to determine
width and colour of gridlines}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Plot the Per Sequence Quality Scores for a set of FASTQC reports
}
\details{
Plots the distribution of average sequence quality scores across the set of files.
Values can be plotted either as counts (\code{counts = TRUE}) or as frequencies (\code{counts = FALSE}).

Any faceting or scale adjustment can be performed after generation of the initial plot,
using the standard methods of ggplot2 as desired.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotSequenceQualities(fdl)

# Also subset the reads to just the R1 files
r1 <- grepl("R1", fileName(fdl))
plotSequenceQualities(fdl[r1])


}
