% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequenceQualitiesHeatmap.R
\name{plotSequenceQualitiesHeatmap}
\alias{plotSequenceQualitiesHeatmap}
\title{Plot the Per Sequence Quality Scores}
\usage{
plotSequenceQualitiesHeatmap(x, subset, labels, counts = FALSE, pwfCols,
  usePlotly = FALSE, dendrogram = FALSE, clusterNames = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{labels}{An optional named factor of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{counts}{\code{logical}. Plot the counts from each file if \code{counts = TRUE}.
If \code{counts = FALSE} the frequencies will be plotted}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} is \code{FALSE}
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{clusterNames}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using heirachial clustering}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}
}
\value{
A ggplot2 object
}
\description{
Plot the Per Sequence Quality Scores for a set of FASTQC reports
}
\details{
Plots the distribution of average sequence quality scores across the set of files.
Values can be plotted either as counts (\code{counts = TRUE}) or as frequencies (\code{counts = FALSE}).

Any faceting or scale adjustment can be performed after generation of the initial plot,
using the standard methods of ggplot2 as desired.
}
