% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwf.R
\docType{data}
\name{pwf}
\alias{pwf}
\title{Colours for PASS/WARN/FAIL}
\format{An object of class \code{PwfCols} of length 1.}
\usage{
pwf
}
\description{
Default colours for PASS/WARN/FAIL
}
\details{
\code{pwf} is an object of class PwfCols supplied with the package and used as the default colouring.
Colours correspond approximately to PASS, WARN and FAIL from the FASTQC reports,
with the additional colour (MAX) included to indicate an extreme FAIL.
In order, these colours in the default vector are green (\code{rgb(0, 0.8, 0)}),
yellow (\code{rgb(0.9, 0.9, 0.2)}), red (\code{rgb(0.8, 0.2, 0.2)}) and white
(\code{rgb(1, 1, 1)})
}
\examples{
# Make a pie chart showing the default colours
plot(pwf)

}
\keyword{datasets}
