% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PwfCols.R
\docType{class}
\name{PwfCols-class}
\alias{PwfCols-class}
\alias{PwfCols}
\title{The PwfCols class and associated methods}
\description{
Define the PwfCols class and associated methods
}
\details{
This is an object of with four colours in components named PASS, WARN, FAIL
and MAX.
Used to indicate these categories as defined on the standard plots from
fastqc.
}
\section{Slots}{

\describe{
\item{\code{PASS}}{A vector of length 1, defining the colour for PASS in rgb format.
Defaults to rgb(0, 0.8, 0)}

\item{\code{WARN}}{A vector of length 1, defining the colour for WARN in rgb format.
Defaults to rgb(0.9, 0.9, 0.2)}

\item{\code{FAIL}}{A vector of length 1, defining the colour for FAIL in rgb format.
Defaults to rgb(0.8, 0.2, 0.2)}

\item{\code{MAX}}{A vector of length 1, defining the colour for an extreme FAIL or
NA in rgb format. Defaults to rgb(1, 1, 1)}
}}

