% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PwfCols.R
\docType{methods}
\name{names,PwfCols-method}
\alias{names,PwfCols-method}
\alias{plot,PwfCols-method}
\title{Define methods for objects of class PwfCols}
\usage{
\S4method{names}{PwfCols}(x)

\S4method{plot}{PwfCols}(x, y, ...)
}
\description{
Define methods for objects of class PwfCols
}
