% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwf.R
\docType{data}
\name{validPwf}
\alias{validPwf}
\alias{pwfCols}
\title{Colours for PASS/WARN/FAIL}
\format{An object of class \code{character} of length 4.}
\usage{
validPwf(x)

pwfCols
}
\arguments{
\item{x}{a vector to be tested as a valid pwfCols vector}

\item{pwfCols}{a default vector supplied with the package. Described in Details below}
}
\value{
validPwf returns a \code{logical} vector of length 1
}
\description{
Default colours for PASS/WARN/FAIL
}
\details{
\code{pwfCols} must be a vector with four colours in rgb format.
Colours correspond to PASS, WARN and FAIL from the FASTQC reports,
with the additional colour denoted as X to indicate an eXtreme FAIL.
In order, these colours in the defulat vector are green (\code{rgb(0, 0.8,0)}),
yellow (\code{rgb(0.9, 0.9, 0.2)}), red (\code{rgb(0.8, 0.2, 0.2)}) and white
(\code{rgb(1, 1, 1)})
}
\keyword{datasets}
