% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHisat2Logs.R
\name{readHisat2Log}
\alias{readHisat2Log}
\title{Read Hisat2 Log Files}
\usage{
readHisat2Log(logs)
}
\arguments{
\item{logs}{\code{character}. Vector of paths to hisat2 log files}
}
\value{
A \code{data_frame}
}
\description{
Import one or more hisat2 log files as a data frame.
}
