% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStarLogs.R
\name{readStarLogs}
\alias{readStarLogs}
\title{Load STAR Alignment Summaries}
\usage{
readStarLogs(path)
}
\arguments{
\item{path}{\code{character}. Path to one or more STAR alignment summaries}
}
\value{
Returns an object of class \code{starLogData}
}
\description{
Loads the information from one or more Log.final.out files as output by STAR
}
