% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastQC.R
\name{runFastQC}
\alias{runFastQC}
\alias{runFastQC,ANY-method}
\title{Deprecated wrapper for the bash shell command fastqc.}
\usage{
runFastQC(object)

\S4method{runFastQC}{ANY}(object)
}
\arguments{
\item{object}{Deprecated}
}
\value{
A warning
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
