% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastQC.R
\name{runFastQC}
\alias{runFastQC}
\title{A wrapper for the bash shell command fastqc.}
\usage{
runFastQC(object, outPath, exec = "/usr/local/bin/fastqc", threads = 1L,
  casava = FALSE, nofilter = FALSE, extract = FALSE, nogroup = FALSE,
  contaminants = c(), adapters = c(), kmers = 7, overWrite = FALSE)
}
\arguments{
\item{object}{A \code{\link{FastqFileList}}}

\item{outPath}{The path to write the FastQC reports. Must exist as it will not be created when calling this function}

\item{exec}{The location of the fastqc executable.}

\item{threads}{The number of threads to run in parallel}

\item{casava}{logical. Sets the \code{--casava} flag}

\item{nofilter}{logical. Sets the \code{--nofilter} flag}

\item{extract}{logical. Extract the zip files on completion of the report}

\item{nogroup}{logical. Sets the grouping of bases for reads longer than 50bp}

\item{contaminants}{Path to an alternate file with contaminants.
The structure of the file will not be checked.
Refer to the \code{fastqc} help page for more details}

\item{adapters}{Path toa file listing adapters to search for.
The structure of the file will not be checked.
Refer to the \code{fastqc} help page for more details}

\item{kmers}{An integer between 2 and 10}

\item{overWrite}{logical. Overwrite any existing reports}
}
\value{
A \code{list} with components
\describe{
\item{fastqcFiles}{The paths to the output files containing the key information. Not the html files}
\item{outPath}{The output path, as specified in the function call}
\item{command}{The command as executed}
\item{version}{The FastQC version number}
}
}
\description{
Takes a FastqFileList
}
\details{
This is a simple wrapper function for controlling & running \code{fastqc} from within R.
This can be very useful for controlling & documenting an entire pipeline from within knitr to produce a simple report

Only the common functinoality of FastQC is implemented,
for more fine detail pease call FastQC directly.
}
\seealso{
\code{\link{FastqFileList}}
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}
