% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamtoolsStats.R
\docType{class}
\name{SamtoolsStats-class}
\alias{SamtoolsStats-class}
\alias{SamtoolsStats}
\title{The SamtoolsStats Object Class}
\usage{
SamtoolsStats(x)
}
\arguments{
\item{x}{Path to a single samtools stats report}
}
\value{
An object of class SamtoolsStats
}
\description{
The SamtoolsStats Object Class
}
\details{
This object class is the main object required for generating plots
and tables. Modules are
contained as individual slots, which can be viewed using \code{slotNames}.

Individual modules can be returned using the function \code{getModule()}
and specifying which module is required. See \code{\link{getModule}} for
more details.

All modules contain tibbles, except for cmd
}
\section{Slots}{

\describe{
\item{\code{Summary_Numbers}}{Summary of alignments}

\item{\code{First_Fragment_Qualities}}{Distribution of quality scores by cycle}

\item{\code{Last_Fragment_Qualities}}{Distribution of quality scores by cycle}

\item{\code{GC_Content_of_first_fragments}}{As above but only for first fragments}

\item{\code{GC_Content_of_last_fragments}}{As above but only for last fragments}

\item{\code{ACGT_content_per_cycle}}{Base content per cycle}

\item{\code{ACGT_content_for_first_fragments}}{As above but only for first fragments}

\item{\code{ACGT_content_for_last_fragments}}{As above but only for last fragments}

\item{\code{Insert_sizes}}{Breakdown of insert sizes and alignments}

\item{\code{Read_lengths}}{Read length distribution}

\item{\code{Read_lengths_first_fragments}}{As above but only for first fragments}

\item{\code{Read_lengths_last_fragments}}{As above but only for last fragments}

\item{\code{Indels_per_cycle}}{Forward and Reverse insertions/deletions per cycle}

\item{\code{Coverage_distribution}}{Coverage by bin}

\item{\code{GC_depth}}{GC distribution with 10th, 25th, 50th, 75th and 90th
percentiles}

\item{\code{Checksums}}{CRC32 Checksums for Read Names, Sequences & Qualities}

\item{\code{cmd}}{The command used for generation of the report}

\item{\code{version}}{The version of samtools & htslib used for generation of the
report}
}}

