% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence_Duplication_Levels.R
\docType{methods}
\name{Sequence_Duplication_Levels,FastqcData-method}
\alias{Sequence_Duplication_Levels,FastqcData-method}
\alias{Sequence_Duplication_Levels}
\alias{Sequence_Duplication_Levels,FastqcDataList-method}
\alias{Sequence_Duplication_Levels,FastqcFile-method}
\alias{Sequence_Duplication_Levels,FastqcFileList-method}
\alias{Sequence_Duplication_Levels,character-method}
\title{Get the Sequence Duplication Levels information}
\usage{
\S4method{Sequence_Duplication_Levels}{FastqcData}(object)

\S4method{Sequence_Duplication_Levels}{FastqcDataList}(object)

\S4method{Sequence_Duplication_Levels}{FastqcFile}(object)

\S4method{Sequence_Duplication_Levels}{FastqcFileList}(object)

\S4method{Sequence_Duplication_Levels}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Sequence Duplication Levels module from one or more FastQC reports
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Print the Sequence Duplication Levels
Sequence_Duplication_Levels(fdl)

}
