% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence_Length_Distribution.R
\docType{methods}
\name{Sequence_Length_Distribution,FastqcData-method}
\alias{Sequence_Length_Distribution,FastqcData-method}
\alias{Sequence_Length_Distribution,FastqcDataList-method}
\alias{Sequence_Length_Distribution,FastqcFile-method}
\alias{Sequence_Length_Distribution,FastqcFileList-method}
\alias{Sequence_Length_Distribution,character-method}
\title{Get the Sequence Length Distribution information}
\usage{
\S4method{Sequence_Length_Distribution}{FastqcData}(object)

\S4method{Sequence_Length_Distribution}{FastqcDataList}(object)

\S4method{Sequence_Length_Distribution}{FastqcFile}(object)

\S4method{Sequence_Length_Distribution}{FastqcFileList}(object)

\S4method{Sequence_Length_Distribution}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files}
}
\value{
A single \code{data_frame} containing all information combined from all supplied FastQC reports
}
\description{
Retrieve the Sequence Length Distribution module from one or more FastQC reports
}
