% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLabels.R
\name{setLabels}
\alias{setLabels}
\title{Perform the checks and return the labels}
\usage{
setLabels(df, labels, pattern = ".(fastq|fq|bam).*")
}
\arguments{
\item{df}{A data.frame with a column titled "Filename"}

\item{labels}{Named vector of labels for plotting}

\item{pattern}{character Regular expression to remove from filenames}
}
\value{
Named character vector
}
\description{
Checks for the presence of labels and returns defaults
}
\details{
Takes a named vector of labels and checks for the correct fields.
If no vector is supplied, returns the file names missing the specified pattern,
which defaults to removing fastq(.gz), fq(.gz) or bam.
}
\examples{
df <- data.frame(Filename = paste0("File1", "File2"), "fastq")
ngsReports:::setLabels(df)

}
\keyword{internal}
