% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColours.R
\docType{methods}
\name{show,PwfCols-method}
\alias{show,PwfCols-method}
\title{The default method for show}
\usage{
\S4method{show}{PwfCols}(object)
}
\arguments{
\item{object}{An object of class PwfCols}
}
\description{
The default method for show
}
