% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcDataList.R
\docType{methods}
\name{[,FastqcDataList,ANY,ANY,ANY-method}
\alias{[,FastqcDataList,ANY,ANY,ANY-method}
\title{Methods for an object of class FastqcDataList}
\usage{
\S4method{[}{FastqcDataList,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{object}{An object of class \code{FastqcDataList}}
}
\value{
Returns a \code{data_frame} for all the methods except \code{path}, \code{names},
\code{fileNames}
}
\description{
Methods and Accessors for an object of class FastqcDataList
}
\details{
Describes all methods for an object of class \code{FastqcDataList}
}
