% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Deduplicated_Percentage.R
\docType{methods}
\name{Total_Deduplicated_Percentage,FastqcData-method}
\alias{Total_Deduplicated_Percentage,FastqcData-method}
\alias{Total_Deduplicated_Percentage}
\alias{Total_Deduplicated_Percentage,FastqcDataList-method}
\alias{Total_Deduplicated_Percentage,FastqcFile-method}
\alias{Total_Deduplicated_Percentage,FastqcFileList-method}
\alias{Total_Deduplicated_Percentage,character-method}
\title{Get the Total Deduplicated Percentage information}
\usage{
\S4method{Total_Deduplicated_Percentage}{FastqcData}(object)

\S4method{Total_Deduplicated_Percentage}{FastqcDataList}(object)

\S4method{Total_Deduplicated_Percentage}{FastqcFile}(object)

\S4method{Total_Deduplicated_Percentage}{FastqcFileList}(object)

\S4method{Total_Deduplicated_Percentage}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{fastqcDataList}, or simply a \code{character}
vector of paths to fastqc files}
}
\value{
A single \code{tibble} containing all information combined from all
supplied FastQC reports
}
\description{
Retrieve the Total Deduplicated Percentage module from one or
more FastQC reports
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Get the estimate total deduplicated percentage
Total_Deduplicated_Percentage(fdl)

}
