% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalGC.R
\docType{methods}
\name{transcriptomes}
\alias{transcriptomes}
\alias{transcriptomes,TheoreticalGC-method}
\title{List Available Transcriptomes}
\usage{
transcriptomes(object)

\S4method{transcriptomes}{TheoreticalGC}(object)
}
\arguments{
\item{object}{An object of class Theoretical GC}
}
\value{
A \code{tibble} object
}
\description{
List available transcriptomes in a TheoreticalGC object
}
\examples{
transcriptomes(gcTheoretical)

}
