% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\docType{methods}
\name{version,FastqcData-method}
\alias{version,FastqcData-method}
\alias{version}
\alias{version,FastqcDataList-method}
\title{Get the FASTQC version}
\usage{
\S4method{version}{FastqcData}(object)

\S4method{version}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class \code{FastqcData} or \code{FastqcDataList}}
}
\value{
A character vector (FastqcData), or tibble (FastqcDataList)
}
\description{
Get the FASTQC version used to generate the intial files
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Get the FASTQC version
version(fdl)

}
