% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHtmlReport.R
\name{writeHtmlReport}
\alias{writeHtmlReport}
\title{Write an HTML Summary Report}
\usage{
writeHtmlReport(fastqcDir, template, targetsDF, overwrite = FALSE,
  quiet = TRUE)
}
\arguments{
\item{fastqcDir}{A directory containing zipped, or extracted FastQC reports}

\item{template}{The template file which will be copied into \code{fastqcDir}}

\item{targetsDF}{A data.frame with at least two columns named \code{Filename} and \code{Label}.
The filenames should match the original fastq files, and the labels should be simply alternative
labels for these files for convenience.}

\item{overwrite}{\code{logical}. Overwrite any previous copies of the template file in the destination directory}

\item{quiet}{\code{logical}. Show or hide markdown output in the Console.}
}
\description{
Compiles an HTML report using a supplied template
}
\details{
This will take a user supplied template, or the file supplied with the package and create an HTML
summary of all standard FASTQC plots for all files in the supplied directory.

Compilation of an html report may require the package \code{webshot}.
This in turn requires an a installation of PhantomJS, which can be installed using \code{webshot::install_phantomjs()}.
If errors are received during compilation, please address this.
}
