% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Statistics.R
\docType{methods}
\name{Basic_Statistics,FastqcData-method}
\alias{Basic_Statistics,FastqcData-method}
\alias{Basic_Statistics}
\alias{Basic_Statistics,FastqcDataList-method}
\alias{Basic_Statistics,FastqcFile-method}
\alias{Basic_Statistics,FastqcFileList-method}
\alias{Basic_Statistics,character-method}
\title{Get the set of Basic Statistics}
\usage{
\S4method{Basic_Statistics}{FastqcData}(object)

\S4method{Basic_Statistics}{FastqcDataList}(object)

\S4method{Basic_Statistics}{FastqcFile}(object)

\S4method{Basic_Statistics}{FastqcFileList}(object)

\S4method{Basic_Statistics}{character}(object)
}
\arguments{
\item{object}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{fastqcDataList}, or simply a \code{character} vector
of paths to fastqc files}
}
\value{
A single \code{tibble} containing all information combined from all
supplied FastQC reports
}
\description{
Retrieve the Basic Statistics module from one or more FastQC
reports
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Return a tibble/tibble with the raw information
Basic_Statistics(fdl)

}
