% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{[,FastqcDataList,numeric,missing-method}
\alias{[,FastqcDataList,numeric,missing-method}
\alias{[,FastqcDataList,character,missing-method}
\alias{[,FastqcDataList,logical,missing-method}
\alias{[,FastqcDataList,ANY,missing-method}
\alias{[,FastpDataList,numeric,missing-method}
\alias{[,FastpDataList,character,missing-method}
\alias{[,FastpDataList,logical,missing-method}
\alias{[,FastpDataList,ANY,missing-method}
\title{Extract Elements}
\usage{
\S4method{[}{FastqcDataList,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastqcDataList,character,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastqcDataList,logical,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastqcDataList,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastpDataList,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastpDataList,character,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastpDataList,logical,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{FastpDataList,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A FastqcDataList or FastpDataList}

\item{i}{character, logical or integer vector}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}
}
\value{
Will return a subset of the original object following the standard
rules for subsetting objects
}
\description{
Extract elements from FastqcDataList Object
}
\details{
Extract elements in a consistent manner with R conventions
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# Subsetting using the standard methods
fdl[1]
fdl[[1]]

}
