% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,FastqcDataList,ANY,missing,ANY-method}
\title{Extract Elements}
\usage{
\S4method{[}{FastqcDataList,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A FastqcDataList}

\item{i}{character, logical or integer vector}

\item{j}{not used}

\item{...}{not used}

\item{drop}{not used}
}
\value{
Will return a subset of the original object following the standard
rules for subsetting objects
}
\description{
Extract elements from FastqcDataList Object
}
\details{
Extract elements in a consistent manner with R conventions
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# Subsetting using the standard methods
fdl[1]
fdl[[1]]

}
