% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcDataList.R
\docType{methods}
\name{path,FastqcDataList-method}
\alias{path,FastqcDataList-method}
\alias{path,FastqcDataList-methods}
\alias{fileNames,FastqcDataList-method}
\title{Methods for an object of class FastqcDataList}
\usage{
\S4method{path}{FastqcDataList}(object)

\S4method{fileNames}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class \code{FastqcDataList}}
}
\value{
Returns a \code{data_frame} for all the methods except \code{path}, \code{names},
\code{fileNames}
}
\description{
Methods and Accessors for an object of class FastqcDataList
}
\details{
Describes all methods for an object of class \code{FastqcDataList}
}
