% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FastqcFile.R
\docType{methods}
\name{FastqcFile}
\alias{FastqcFile}
\alias{FastqcFile,character-method}
\alias{FastqcFileList}
\alias{FastqcFileList,character-method}
\alias{FastqcFileList,character-method}
\title{Path to data from FastQC}
\usage{
FastqcFile(path)

\S4method{FastqcFile}{character}(path)

FastqcFileList(path)

\S4method{FastqcFileList}{character}(path)
}
\arguments{
\item{path}{The path to a single FastQC output zip file, or uncompressed folder}
}
\value{
An object of cass \code{FastqcFile}
}
\description{
Define a FastqcFile
}
\details{
Checks the structure of a folder or zip file with the output from FastQC
}
