% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastqcData.R, R/methods-FastqcFile.R
\docType{methods}
\name{getFastqcData,character-method}
\alias{getFastqcData,character-method}
\alias{getFastqcData,character-methods}
\alias{FastqcFile,character-method}
\alias{path,FastqcFile-method}
\alias{isCompressed,FastqcFile-method}
\alias{fileNames,FastqcFile-method}
\alias{show,FastqcFile-method}
\alias{show,FastqcFile-methods}
\title{Path to FastQC data from a single fastq file}
\usage{
\S4method{getFastqcData}{character}(object)

\S4method{FastqcFile}{character}(filePath)

\S4method{path}{FastqcFile}(object)

\S4method{isCompressed}{FastqcFile}(object)

\S4method{fileNames}{FastqcFile}(object)

\S4method{show}{FastqcFile}(object)
}
\arguments{
\item{object}{An object of class \code{FastqcFile}}

\item{filePath}{The path to a single FastQC output zip file, or uncompressed folder}
}
\value{
An object of cass \code{FastqcFile}
}
\description{
Define a FastqcFile
}
\details{
Checks the structure of a folder or zip file with the output from FastQC
}
