% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{FastqcFile}
\alias{FastqcFile}
\title{Define object classes}
\usage{
FastqcFile(path)
}
\description{
Define object classes
}
\section{Slots}{

\describe{
\item{\code{path}}{a character representation of a path to a single file/directory}

\item{\code{compressed}}{logical. Is the FastQC file compressed as a zip file}
}}

