% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFile.R
\docType{methods}
\name{FastqcFile}
\alias{FastqcFile}
\title{Create a New FastqcFile Object}
\usage{
FastqcFile(x)
}
\arguments{
\item{x}{Character vector (1) specifying a valid path to a file/directory as output by FastQC}
}
\description{
Create a FastqcFile object
}
