% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFile.R
\docType{class}
\name{FastqcFile-class}
\alias{FastqcFile-class}
\alias{FastqcFile}
\alias{FastqcFile,character-method}
\alias{FastqcFile}
\title{The FastqcFile Object Class}
\usage{
FastqcFile(x)

\S4method{FastqcFile}{character}(x)
}
\arguments{
\item{x}{Character vector (1) specifying a valid path to a file/directory as output by FastQC}
}
\value{
An object of class FastqcFile
}
\description{
The FastqcFile Object Class
}
\details{
The is an object which refers to a fastqc output file.
Only the path is stored, however the file is checked for the correct structure on formation of the object.
Files can be zipped (*_fastqc.zip) or extracted directories
}
\section{Slots}{

\describe{
\item{\code{path}}{Character vector of length 1 which contains a valid file path.}
}}

\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# As this is the root structure, we can only call this 
# function with an individual file
ff <- FastqcFile(fileList[[1]])

}
