% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFile.R
\docType{methods}
\name{FastqcFile,character-method}
\alias{FastqcFile,character-method}
\alias{FastqcFile}
\title{Create a new FastqcFile Object}
\usage{
\S4method{FastqcFile}{character}(x)
}
\arguments{
\item{x}{Character vector (1) specifying a valid path to a file/directory as output by FastQC}
}
\value{
An object of class FastqcFile
}
\description{
Create a new FastqcFile Object
}
\details{
Create a new FastqcFile Object from an external file
}
