% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastqcData.R, R/methods-FastqcFileList.R
\docType{methods}
\name{getFastqcData,FastqcFileList-method}
\alias{getFastqcData,FastqcFileList-method}
\alias{getFastqcData,FastqcFileList-methods}
\alias{FastqcFileList,character-method}
\alias{FastqcFileList,list-method}
\alias{path,FastqcFileList-method}
\alias{isCompressed,FastqcFileList-method}
\alias{names,FastqcFileList-method}
\title{Path to FastQC data a vector of fastq files}
\usage{
\S4method{getFastqcData}{FastqcFileList}(object)

\S4method{FastqcFileList}{character}(path)

\S4method{FastqcFileList}{list}(path)

\S4method{path}{FastqcFileList}(object)

\S4method{isCompressed}{FastqcFileList}(object)

\S4method{names}{FastqcFileList}(x)
}
\arguments{
\item{object}{An object of class \code{FastqcFileList}}

\item{path}{The path to a vector of zip files (or uncompressed folders) as output by FastQC}
}
\description{
Define a FastqcFileList
}
\details{
Checks the structure of a folder or zip file with the output from FastQC
}
