% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFileList.R
\docType{methods}
\name{FastqcFileList}
\alias{FastqcFileList}
\alias{FastqcFileList,list-method}
\alias{FastqcFileList}
\title{Create a new FastqcFileList Object}
\usage{
FastqcFileList(path)

\S4method{FastqcFileList}{list}(path)
}
\arguments{
\item{x}{Character vector specifying a valid paths to files/directories as output by FastQC}
}
\value{
An object of class FastqcFileList
}
\description{
Create a new FastqcFileList Object
}
\details{
Create a new FastqcFileList Object from a vector of file paths
}
