% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFileList.R
\docType{class}
\name{FastqcFileList-class}
\alias{FastqcFileList-class}
\alias{FastqcFileList}
\alias{FastqcFileList,character-method}
\alias{FastqcFileList}
\alias{FastqcFileList,list-method}
\alias{FastqcFileList}
\title{The FastqcFileList Object Class}
\usage{
FastqcFileList(x)

\S4method{FastqcFileList}{character}(x)

\S4method{FastqcFileList}{list}(x)
}
\arguments{
\item{x}{Character vector specifying a valid paths to files/directories as output by FastQC}
}
\value{
An object of class FastqcFile
}
\description{
The FastqcFileList Object Class
}
\details{
The is an object which refers to a list of fastqc output files.
Only the paths are stored, however the files are checked for the correct structure on formation of the object.
Files can be any combination of zipped (*_fastqc.zip) or extracted directories
}
\section{Slots}{

\describe{
\item{\code{...}}{this can either be a single character vector of paths to FASTQC files, or several instances of FastqcFile objects}
}}

\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcFileList
ffl <- FastqcFileList(fileList)

}
