% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqcFileList.R
\docType{class}
\name{FastqcFileList-class}
\alias{FastqcFileList-class}
\alias{FastqcFileList}
\alias{FastqcFileList,character-method}
\alias{FastqcFileList,list-method}
\title{The FastqcFileList Object Class}
\usage{
FastqcFileList(x)

\S4method{FastqcFileList}{character}(x)

\S4method{FastqcFileList}{list}(x)
}
\arguments{
\item{x}{Character vector specifying a valid paths to files/directories as
output by FastQC}
}
\value{
An object of class FastqcFile
}
\description{
The FastqcFileList Object Class
}
\details{
The is an object which refers to a list of fastqc output files.
Only the paths are stored, however the files are checked for the correct
structure on formation of the object.
Files can be any combination of zipped (*_fastqc.zip) or extracted
directories
}
\section{Slots}{

\describe{
\item{\code{...}}{this can either be a single character vector of paths to FASTQC
files, or several instances of FastqcFile objects}
}}

\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcFileList
ffl <- FastqcFileList(fileList)

}
