% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastQC_shinyApp.R
\name{fastqcShiny}
\alias{fastqcShiny}
\title{Run fastQC shiny app}
\usage{
fastqcShiny(fastqcInput = NULL)
}
\arguments{
\item{fastqcInput}{can be a \code{FastqcFileList}, \code{fastqcDataList},
or simply a \code{character} vector of paths to fastqc files.}
}
\value{
UI data for fastQC shiny.
}
\description{
Returns a shiny app interface to parse many fastQC objects
}
\details{
Currently some plots can take a while to render if the \code{FastqcDataList} passed to
\code{fastqcInput} has many elements
}
\examples{
\dontrun{
#' # Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Run the Shiny app
fastqcShiny(fdl)
}

}
