% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileName.R
\docType{methods}
\name{fileName}
\alias{fileName}
\alias{fileName,FastqcFile-method}
\alias{fileName,FastqcFileList-method}
\alias{fileName,FastqcData-method}
\alias{fileName,FastqcDataList-method}
\title{Return the File Names from an Object}
\usage{
\S4method{fileName}{FastqcFile}(object)

\S4method{fileName}{FastqcFileList}(object)

\S4method{fileName}{FastqcData}(object)

\S4method{fileName}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class FastqcFile, FastqcFileList, FastqcData or 
FastqcDataList}
}
\value{
Returns the filenames, without the preceding directories, i.e. basename.
For a FastqcFile/FastqcFileList these will be the underlying FastQC files.
For a FastqcData/FastqcDataList, these will correspond to the fastq files 
about which the FastQC report was written.
}
\description{
The File Names from an object
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)

# Form a FastqcFileList
ffl <- FastqcFileList(fileList)
fileName(ffl)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)
fileName(fdl)

}
