% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqName.R
\docType{methods}
\name{fqName}
\alias{fqName}
\alias{fqName,ANY-method}
\alias{fqName,FastqcData-method}
\alias{fqName,FastqcDataList-method}
\title{Return the Underlying Fastq File Names from FastqcData* Objects}
\usage{
fqName(object)

\S4method{fqName}{ANY}(object)

\S4method{fqName}{FastqcData}(object)

\S4method{fqName}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class FastqcData or FastqcDataList}
}
\value{
Returns the names of the Fastq files the FastQC report was
generated from, without any preceding directories.
}
\description{
Return the Underlying Fastq File Names from FastqcData* Objects
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)
fqName(fdl)

}
