% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PwfCols.R
\docType{methods}
\name{getColours}
\alias{getColours}
\alias{getColours,PwfCols-method}
\alias{getColors}
\alias{setColours,PwfCols-method}
\alias{setColours}
\alias{setColors}
\alias{setAlpha,PwfCols-method}
\alias{setAlpha}
\title{Define the generics & methods for setting/getting colours from an object of class PwfCols}
\usage{
getColours(x)

\S4method{getColours}{PwfCols}(x)

getColours(x)

\S4method{setColours}{PwfCols}(x, PASS, WARN, FAIL, MAX)

setColours(x, ...)

\S4method{setAlpha}{PwfCols}(x, alpha)
}
\arguments{
\item{x}{An object of class PwfCols}
}
\description{
Define the generics & methods for setting/getting colours from an object of class PwfCols

Methods to return (get) the colours from a PwfCols object

Methods to set the colours for a PwfCols object

Define the generic and method for setting alpha (transparency)
}
