% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColours.R
\docType{methods}
\name{getColours,PwfCols-method}
\alias{getColours,PwfCols-method}
\alias{setColours,PwfCols-method}
\alias{setColours}
\title{Get or set the colours in a PwfCols object}
\usage{
\S4method{getColours}{PwfCols}(object)

\S4method{setColours}{PwfCols}(object, PASS, WARN, FAIL, MAX)
}
\arguments{
\item{object}{An object of class PwfCols}

\item{PASS}{The colour used to signify a PASS}

\item{WARN}{The colour used to signify a WARN}

\item{FAIL}{The colour used to signify a FAIL}

\item{MAX}{The colour used as an extreme FAIL}
}
\value{
An object of class \code{PwfCols}
}
\description{
Get or set the colours in a PwfCols object
}
\details{
All colours must be specified in rgb format.
}
