% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastqcData.R
\docType{methods}
\name{getFastqcData}
\alias{getFastqcData}
\alias{getFastqcData,FastqcFile-method}
\alias{getFastqcData,NULL-method}
\alias{getFastqcData,character-method}
\alias{getFastqcData,FastqcFileList-method}
\title{Get all data from FastQC files}
\usage{
getFastqcData(object)

\S4method{getFastqcData}{FastqcFile}(object)

\S4method{getFastqcData}{`NULL`}(object)

\S4method{getFastqcData}{character}(object)

\S4method{getFastqcData}{FastqcFileList}(object)
}
\arguments{
\item{object}{Can be a FastqcFile or FastqcFileList, or paths to files}
}
\value{
An object of \code{FastqcData} or a \code{FastqcDataList}
}
\description{
Read the information from the \code{fastqc_data.txt} files in each FastqcFile
}
\examples{
# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")
# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

}
