% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFastqcData.R
\docType{methods}
\name{getFastqcData}
\alias{getFastqcData}
\alias{getFastqcData}
\alias{getFastqcData,FastqcFile-method}
\alias{getFastqcData}
\alias{getFastqcData,character-method}
\alias{getFastqcData}
\alias{getFastqcData,FastqcFileList-method}
\title{Get all data from FastQC files}
\usage{
getFastqcData(object)

\S4method{getFastqcData}{FastqcFile}(object)

\S4method{getFastqcData}{character}(object)

\S4method{getFastqcData}{FastqcFileList}(object)
}
\arguments{
\item{object}{Can be a FastqcFile or FastqcFileList, or paths to files}
}
\value{
An object of \code{FastqcData} or a \code{FastqcDataList}
}
\description{
Read the information from the \code{fastqc_data.txt} files in each FastqcFile
}
