% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\docType{methods}
\name{getSummary,FastqcFile-method}
\alias{getSummary,FastqcFile-method}
\alias{getSummary}
\alias{getSummary,FastqcFileList-method}
\alias{getSummary,character-method}
\alias{getSummary,FastqcData-method}
\alias{getSummary,FastqcDataList-method}
\title{Get the summary information from Fastqc Files}
\usage{
\S4method{getSummary}{FastqcFile}(object)

\S4method{getSummary}{FastqcFileList}(object)

\S4method{getSummary}{character}(object)

\S4method{getSummary}{FastqcData}(object)

\S4method{getSummary}{FastqcDataList}(object)
}
\arguments{
\item{object}{Can be a FastqcFile or FastqcFileList}
}
\value{
A \code{tibble} will be returned when supplying a \code{FastqcFile} 
object, whilst a list of tibbles will be returned when supplying a 
\code{FastqcFileList} object
}
\description{
Read the information from the \code{summary.txt} files in each 
FastqcFile
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Return a tibble/tibble with the raw information
getSummary(fdl)

}
