% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\docType{methods}
\name{getSummary,FastqcFile-method}
\alias{getSummary,FastqcFile-method}
\alias{getSummary}
\alias{getSummary,FastqcFileList-method}
\alias{getSummary}
\alias{getSummary,character-method}
\alias{getSummary}
\alias{getSummary,FastqcData-method}
\alias{getSummary}
\alias{getSummary,FastqcDataList-method}
\alias{getSummary}
\title{Get the summary information from Fastqc Files}
\usage{
\S4method{getSummary}{FastqcFile}(object)

\S4method{getSummary}{FastqcFileList}(object)

\S4method{getSummary}{character}(object)

\S4method{getSummary}{FastqcData}(object)

\S4method{getSummary}{FastqcDataList}(object)
}
\arguments{
\item{object}{Can be a FastqcFile or FastqcFileList}
}
\value{
A \code{tibble} will be returned when supplying a \code{FastqcFile} object,
whilst a list of tibbles will be returned when supplying a \code{FastqcFileList} object
}
\description{
Read the information from the \code{summary.txt} files in each FastqcFile
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Return a data_frame/tibble with the raw information
getSummary(fdl)

}
