% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDuplicationMetrics.R
\name{importDuplicationMetrics}
\alias{importDuplicationMetrics}
\title{Import Metrics From MarkDuplicates}
\usage{
importDuplicationMetrics(x)
}
\arguments{
\item{x}{\code{character}. Vector of filenames}
}
\value{
A list with two tibbles.
The first (\code{$metrics}) is the metrics for each file, with the second (\code{$histogram}) being the histogram for each file
}
\description{
Imports metrics files as generated by MarkDuplicates
}
\details{
Files generated by MarkDuplicates from the picard set of tools can be imported using this tool.
Files must be exactly as generated by this tool, without modification
}
