% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStarLogs.R
\name{importStarLogs}
\alias{importStarLogs}
\title{Import STAR log files}
\usage{
importStarLogs(path)
}
\arguments{
\item{path}{\code{character}. Vector of filenames}
}
\value{
A \code{data_frame}.
Column names are broadly similar to those in the supplied files,
but have been modified for easier handling under R naming conventions.

In addition, the columns \code{Mapping_Duration} and the overall \code{Mapping_Rate}
are returned
}
\description{
Imports \code{Log.final.out} files generated by STAR
}
\details{
Imports one or more \code{Log.final.out} files as output by the aligner STAR
Values which are simple to calculate from the returned information are returned
as proportions intead of percentages.
Indel and substitution rates are left as character values showing percentages.
}
