% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSidebar.R
\name{makeSidebar}
\alias{makeSidebar}
\title{Shortcut for making the status sidebar}
\usage{
makeSidebar(status, key, pwfCols)
}
\arguments{
\item{status}{A data.frame with columns 'Filename' & 'Status'}

\item{key}{A vector of values corresponding to the Filename column}

\item{pwfCols}{An object of class PwfCols}
}
\value{
A plotly object. The sidebar for an interactive plot showing PASS/WARN/FAIL 
status for each file.
}
\description{
Shortcut for making the status sidebar
}
\keyword{internal}
