% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdapterContent.R
\docType{methods}
\name{plotAdapterContent}
\alias{plotAdapterContent}
\alias{plotAdapterContent,character-method}
\alias{plotAdapterContent,character}
\alias{plotAdapterContent,FastqcFile-method}
\alias{plotAdapterContent,FastqcFile}
\alias{plotAdapterContent,FastqcFileList-method}
\alias{plotAdapterContent,FastqcFileList}
\alias{plotAdapterContent,FastqcData-method}
\alias{plotAdapterContent,FastqcData}
\alias{plotAdapterContent,FastqcDataList-method}
\alias{plotAdapterContent,FastqcDataList}
\title{Draw an Adapter Content Plot}
\usage{
plotAdapterContent(x, usePlotly = FALSE, ...)

\S4method{plotAdapterContent}{character}(x, usePlotly = FALSE, ...)

\S4method{plotAdapterContent}{FastqcFile}(x, usePlotly = FALSE, ...)

\S4method{plotAdapterContent}{FastqcFileList}(x, usePlotly = FALSE, ...)

\S4method{plotAdapterContent}{FastqcData}(x, usePlotly = FALSE, pwfCols,
  warn = 5, fail = 10, ...)

\S4method{plotAdapterContent}{FastqcDataList}(x, usePlotly = FALSE,
  plotType = "heatmap", labels, adapterType = "Total", pwfCols, warn = 5,
  fail = 10, clusterNames = FALSE, dendrogram = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. Output as ggplot2 (default) or plotly object.}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{warn, fail}{The default values for warn and fail are 5 and 10 respectively (i.e. precentages)}

\item{plotType}{\code{character}. Can only take the values \code{plotType = "heatmap"}
or \code{plotType = "line"}}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{adapterType}{A regular expression matching the adapter(s) to be plotted.
To plot all adapters summed, specify \code{adapterType = "Total"}.
This is the default behaviour.}

\item{clusterNames}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} is \code{FALSE}
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Draw an Adapter Content Plot across one or more FASTQC reports
}
\details{
This extracts the Adapter_Content from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.

When \code{x} is a single FastqcFile, or FastqcData object line plots will always
be drawn for all adapters.
Otherwise, users can select line plots or heatmaps.
When plotting more than one fastqc file, any undetected adapters will not be shown.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotAdapterContent(fdl)

# Also subset the reads to just the R1 files
r1 <- grepl("R1", fileName(fdl))
plotAdapterContent(fdl[r1])

# Plot just the Universal Adapter
# and change the y-axis using ggplot2::scale_y_continuous
library(ggplot2)
plotAdapterContent(fdl, adapterType ="Universal", plotType = "line") +
scale_y_continuous()

}
