% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdapterContent.R
\docType{methods}
\name{plotAdapterContent}
\alias{plotAdapterContent}
\alias{plotAdapterContent,character-method}
\alias{plotAdapterContent,character}
\alias{plotAdapterContent,FastqcFile-method}
\alias{plotAdapterContent,FastqcFile}
\alias{plotAdapterContent,FastqcFileList-method}
\alias{plotAdapterContent,FastqcFileList}
\alias{plotAdapterContent,FastqcData-method}
\alias{plotAdapterContent,FastqcData}
\alias{plotAdapterContent,FastqcDataList-method}
\alias{plotAdapterContent,FastqcDataList}
\title{Draw an Adapter Content Plot}
\usage{
plotAdapterContent(x, usePlotly = FALSE, labels, pwfCols, warn = 5,
  fail = 10, ...)

\S4method{plotAdapterContent}{character}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 5, fail = 10, ...)

\S4method{plotAdapterContent}{FastqcFile}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 5, fail = 10, ...)

\S4method{plotAdapterContent}{FastqcFileList}(x, usePlotly = FALSE,
  labels, pwfCols, warn = 5, fail = 10, ...)

\S4method{plotAdapterContent}{FastqcData}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 5, fail = 10, ...)

\S4method{plotAdapterContent}{FastqcDataList}(x, usePlotly = FALSE,
  labels, pwfCols, warn = 5, fail = 10, plotType = c("heatmap",
  "line"), adapterType = "Total", cluster = FALSE,
  dendrogram = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. Output as ggplot2 (default) or plotly object.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours 
for PASS/WARN/FAIL}

\item{warn, fail}{The default values for warn and fail are 5 and 10 respectively 
(i.e. percentages)}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{plotType}{\code{character}. Can only take the values 
\code{plotType = "heatmap"} or \code{plotType = "line"}}

\item{adapterType}{A regular expression matching the adapter(s) to be plotted.
To plot all adapters summed, specify \code{adapterType = "Total"}.
This is the default behaviour.}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE} 
then the dendrogram will be displayed.}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Draw an Adapter Content Plot across one or more FASTQC reports
}
\details{
This extracts the Adapter_Content from the supplied object and generates a 
ggplot2 object, with a set of minimal defaults.
The output of this function can be further modified using the standard 
ggplot2 methods.

When \code{x} is a single FastqcFile, or FastqcData object line plots will 
always be drawn for all adapters.
Otherwise, users can select line plots or heatmaps.
When plotting more than one fastqc file, any undetected adapters will not be shown.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# The default plot
plotAdapterContent(fdl)

# Also subset the reads to just the R1 files
r1 <- grepl("R1", fileName(fdl))
plotAdapterContent(fdl[r1])

# Plot just the Universal Adapter
# and change the y-axis using ggplot2::scale_y_continuous
plotAdapterContent(fdl, adapterType ="Universal", plotType = "line") +
facet_wrap(~Filename) +
guides(colour = FALSE)

}
