% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdapterContent.R
\name{plotAdapterContent}
\alias{plotAdapterContent}
\title{Draw an Adapter Content Plot}
\usage{
plotAdapterContent(x, subset, facet = TRUE, nc = 2, ylim = c(0, 100),
  trimNames = TRUE, pattern = "(.+)\\\\.(fastq|fq).*", type, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{facet}{logical. Determines whether to use \code{facet_wrap()} or not}

\item{nc}{Number of columns to use when faceting by Adapter Type.}

\item{ylim}{A \code{numeric vector} providing limits for the y-axis.
Defaults to \code{ylim = c(0, 100)}}

\item{trimNames}{\code{logical}. Remove the file suffix from the names displyed in the legend.}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{type}{A regular expression used to filter which adapter(s) are plotted}

\item{...}{Used to pass additional arguments, such as \code{invert} or \code{ignore.case},
to \code{grep} when selecting Adapter \code{type}}
}
\value{
A standard ggplot2 object
}
\description{
Draw an Adapter Content Plot across one or more FASTQC reports
}
\details{
This extracts the Adapter_Content from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.

To set any faceting parameters outside the function, such as \code{scales = "free_y"},
set \code{facet = FALSE} and apply \code{facet_wrap()} after the call to \code{plotAdapterContent}.
Note that the only column available for faceting with be \code{Type},
so the formula \code{~Type} will still need to be applied.

Preset axis limits can also be overwritten easily by adding a call to \code{scale_y_continuous()}
after the call to \code{plotAdapterContent}.
}
