% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdapterContent.R
\name{plotAdapterContent}
\alias{plotAdapterContent}
\title{Draw an Adapter Content Plot}
\usage{
plotAdapterContent(x, subset, adapterType, plotType = "heatmap", warn = 5,
  fail = 10, pwfCols, labels)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{adapterType}{A regular expression used to filter which adapter(s) are plotted}

\item{plotType}{\code{character}. Can only take the values \code{plotType = "heatmap"}
or \code{plotType = "line"}}

\item{warn, fail}{The default values for warn and fail are 5 and 10 respectively (i.e. precentages)}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}
}
\value{
A standard ggplot2 object
}
\description{
Draw an Adapter Content Plot across one or more FASTQC reports
}
\details{
This extracts the Adapter_Content from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.


Preset axis limits can also be overwritten easily by adding a call to \code{scale_y_continuous}
after the call to \code{plotAdapterContent}.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotAdapterContent(fdl)

# Also subset the reads to just the R1 files
r1 <- grepl("R1", fileName(fdl))
plotAdapterContent(fdl, subset = r1)

# Plot just the Universal Adapter
# and change the y-axis using ggplot2::scale_y_continuous
library(ggplot2)
plotAdapterContent(fdl, adapterType ="Universal", plotType = "line") +
scale_y_continuous()

}
