% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAdapterContent.R
\name{plotAdapterContent}
\alias{plotAdapterContent}
\title{Draw an Adapter Content Plot}
\usage{
plotAdapterContent(x, subset, adapterType, plotType = "heatmap", nc, pwfCols,
  trimNames = TRUE, pattern = "(.+)\\\\.(fastq|fq).*", ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{adapterType}{A regular expression used to filter which adapter(s) are plotted}

\item{plotType}{\code{character}. Can only take the values \code{plotType = "heatmap"}
or \code{plotType = "line"}}

\item{nc}{Number of columns to use when faceting by Adapter Type.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{trimNames}{\code{logical}. Remove the file suffix from the names displyed in the legend.}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{...}{Used to pass additional arguments, such as \code{invert} or \code{ignore.case},
to \code{grep} when selecting Adapter \code{adapterType}}
}
\value{
A standard ggplot2 object
}
\description{
Draw an Adapter Content Plot across one or more FASTQC reports
}
\details{
This extracts the Adapter_Content from the supplied object and generates a ggplot2 object,
with a set of minimal defaults.
The output of this function can be further modified using the standard ggplot2 methods.


Preset axis limits can also be overwritten easily by adding a call to \code{scale_y_continuous}
after the call to \code{plotAdapterContent}.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "fastqcReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotAdapterContent(fdl)

# Also subset the reads to just the R1 files
r1 <- grepl("R1", fileNames(fdl))
plotAdapterContent(fdl, subset = r1)

# Plot just the Universal Adapter, and change the y-axis
plotAdapterContent(fdl, adapterType ="Universal", plotType = "line") +
scale_y_continuous()


}
