% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseQualities.R
\docType{methods}
\name{plotBaseQualities}
\alias{plotBaseQualities}
\alias{plotBaseQualities,character-method}
\alias{plotBaseQualities,character}
\alias{plotBaseQualities,FastqcFile-method}
\alias{plotBaseQualities,FastqcFile}
\alias{plotBaseQualities,FastqcFileList-method}
\alias{plotBaseQualities,FastqcFileList}
\alias{plotBaseQualities,FastqcData-method}
\alias{plotBaseQualities,FastqcData}
\alias{plotBaseQualities,FastqcDataList-method}
\alias{plotBaseQualities,FastqcDataList}
\title{Plot the Base Qualities for each file}
\usage{
plotBaseQualities(x, usePlotly = FALSE, ...)

\S4method{plotBaseQualities}{character}(x, usePlotly = FALSE, ...)

\S4method{plotBaseQualities}{FastqcFile}(x, usePlotly = FALSE, ...)

\S4method{plotBaseQualities}{FastqcFileList}(x, usePlotly = FALSE, ...)

\S4method{plotBaseQualities}{FastqcData}(x, usePlotly = FALSE, pwfCols,
  warn = 25, fail = 20, ...)

\S4method{plotBaseQualities}{FastqcDataList}(x, usePlotly = FALSE,
  plotType = c("heatmap", "boxplot"), plotValue = "Mean",
  cluster = FALSE, labels, dendrogram = FALSE, nc = 2, pwfCols,
  warn = 25, fail = 20, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for pass, warning, and fail
values in plot}

\item{warn, fail}{The default values for warn and fail are 30 and 20 respectively (i.e. percentages)}

\item{plotType}{\code{character} Can be either \code{"boxplot"} or \code{"heatmap"}}

\item{plotValue}{\code{character} Type of quality data to be presented "Mean" or "Median"}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{nc}{\code{numeric}. The number of columns to create in the plot layout}
}
\value{
A standard ggplot2 object
}
\description{
Plot the Base Qualities for each file as separate plots
}
\details{
This replicates the \code{Per base sequence quality} plots from FASTQC,
using facets to plce them all in a single ggplot2 object.

For large datasets, subsetting by R1 or R2 reads may be helpful
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot for multiple libraries is a heatmap
plotBaseQualities(fdl)

# The default plot for a single library is the standard boxplot
plotBaseQualities(fdl[[1]])

}
