% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseQualities.R
\name{plotBaseQualities}
\alias{plotBaseQualities}
\title{Plot the Base Qualities for each file}
\usage{
plotBaseQualities(x, subset, nc = 2, pwfCols, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{nc}{\code{numeric}. The number of columns to create in the plot layout}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot the Base Qualities for each file as separate plots
}
\details{
This replicates the \code{Per base sequence quality} plots from FASTQC,
using facets to plce them all in a single ggplot2 object.

For large datasets, subsetting by R1 or R2 reads may be helpful
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default and subset plot
plotBaseQualities(fdl)
r1 <- grepl("R1", fileNames(fdl))
plotBaseQualities(fdl, subset = r1 )

}
