% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseQuals.R
\docType{methods}
\name{plotBaseQuals}
\alias{plotBaseQuals}
\alias{plotBaseQuals,ANY-method}
\alias{plotBaseQuals,FastqcData-method}
\alias{plotBaseQuals,FastqcDataList-method}
\alias{plotBaseQuals,FastpDataList-method}
\title{Plot the Base Qualities for each file}
\usage{
plotBaseQuals(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 25,
  fail = 20,
  boxWidth = 0.8,
  ...
)

\S4method{plotBaseQuals}{ANY}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 25,
  fail = 20,
  boxWidth = 0.8,
  ...
)

\S4method{plotBaseQuals}{FastqcData}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 25,
  fail = 20,
  boxWidth = 0.8,
  ...
)

\S4method{plotBaseQuals}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 25,
  fail = 20,
  boxWidth = 0.8,
  plotType = c("heatmap", "boxplot"),
  plotValue = "Mean",
  cluster = FALSE,
  dendrogram = FALSE,
  nc = 2,
  heat_w = 8L,
  ...
)

\S4method{plotBaseQuals}{FastpDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 25,
  fail = 20,
  plotType = c("heatmap"),
  reads = c("read1", "read2"),
  module = c("Before_filtering", "After_filtering"),
  plotValue = c("mean", "A", "T", "C", "G"),
  cluster = FALSE,
  dendrogram = FALSE,
  heat_w = 8L,
  ...
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or character
vector of file paths}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link[=PwfCols]{PwfCols()}} to give colours for
pass, warning, and fail values in plot}

\item{warn, fail}{The default values for warn and fail are 30 and 20
respectively (i.e. percentages)}

\item{boxWidth}{set the width of boxes when using a boxplot}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{plotType}{\code{character} Can be either \code{"boxplot"} or
\code{"heatmap"}}

\item{plotValue}{\code{character} Type of data to be presented. Can be
any of the columns returned by the appropriate call to \code{getModule()}}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{nc}{\code{numeric}. The number of columns to create in the plot layout.
Only used if drawing boxplots for multiple files in a FastqcDataList}

\item{heat_w}{Relative width of any heatmap plot components}

\item{reads}{Plot heatmaps for read1 or read2 when using a FastpDataList}

\item{module}{Select Before and After filtering when using a FastpDataList}
}
\value{
A standard ggplot2 object or an interactive plotly object
}
\description{
Plot the Base Qualities for each file as separate plots
}
\details{
When acting on a \code{FastqcDataList}, this defaults to a heatmap
using the mean Per_base_sequence_quality score. A set of plots which
replicate those obtained through a standard FastQC html report can be
obtained by setting \code{plotType = "boxplot"}, which uses \code{facet_wrap}
to provide the layout as a single ggplot object.

When acting an a \code{FastqcData} object, this replicates the
\verb{Per base sequence quality} plots from FastQC with no faceting.

For large datasets, subsetting by R1 or R2 reads may be helpful.

An interactive plot can be obtained by setting \code{usePlotly = TRUE}.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# The default plot for multiple libraries is a heatmap
plotBaseQuals(fdl)

# The default plot for a single library is the standard boxplot
plotBaseQuals(fdl[[1]])

}
