% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombinedBaseQualities.R
\name{plotCombinedBaseQualities}
\alias{plotCombinedBaseQualities}
\title{Plot the combined Per_base_sequence_quality information}
\usage{
plotCombinedBaseQualities(x, subset, value = "Mean", pwfCols,
  trimNames = TRUE, pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{value}{\code{character}. Specify which value whould be plotted.
Can be any of the columns returned by \code{\link{Per_base_sequence_quality}}.
Defaults to \code{value = "Mean"}.
Can additionally set to "all" to plot all available quantities}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot the Per_base_sequence_quality information for a set of FASTQC reports
}
\details{
This enables plotting of any of the supplied values (i.e. Mean, Median etc) across
a set of FASTQC reports.
By default only the Mean will be plotted,
however any number of the supplied values can be added to the plot,
and these will be separated by linetype.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Find the R1 files
r1 <- grepl("R1", fileNames(fdl))

# The default plot using the Mean only
plotCombinedBaseQualities(fdl)

# Plot the R1 files showing the Mean and Lower_Quartile
plotCombinedBaseQualities(fdl, subset = r1, value = c("Mean", "Lower_Quartile"))

}
