% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDupLevels.R
\docType{methods}
\name{plotDupLevels}
\alias{plotDupLevels}
\alias{plotDupLevels,ANY-method}
\alias{plotDupLevels,character-method}
\alias{plotDupLevels,FastqcData-method}
\alias{plotDupLevels,FastqcDataList-method}
\title{Plot the combined Sequence_Duplication_Levels information}
\usage{
plotDupLevels(x, usePlotly = FALSE, labels, pwfCols, ...)

\S4method{plotDupLevels}{ANY}(x, usePlotly = FALSE, labels, pwfCols, ...)

\S4method{plotDupLevels}{character}(x, usePlotly = FALSE, labels, pwfCols, ...)

\S4method{plotDupLevels}{FastqcData}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 20,
  fail = 50,
  lineCols = c("red", "blue"),
  ...
)

\S4method{plotDupLevels}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 20,
  fail = 50,
  deduplication = c("pre", "post"),
  plotType = c("heatmap", "line"),
  cluster = FALSE,
  dendrogram = FALSE,
  heatCol = inferno(50),
  ...
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for
pass, warning, and fail values in the plot}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{warn, fail}{The default values for warn and fail are 20 and 50
respectively (i.e. percentages)}

\item{lineCols}{Colours of the lines drawn for individual libraries}

\item{deduplication}{Plot Duplication levels 'pre' or 'post' deduplication.
Can only take values "pre" and "post"}

\item{plotType}{Choose between "heatmap" and "line"}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{heatCol}{Colour palette used for the heatmap}
}
\value{
A standard ggplot2 or plotly object
}
\description{
Plot the Sequence_Duplication_Levels information for a set of
FASTQC reports
}
\details{
This extracts the Sequence_Duplication_Levels from the supplied object and
generates a ggplot2 object, with a set of minimal defaults. For multiple
reports, this defaults to a heatmap with block sizes proportional to the
percentage of reads belonging to that duplication category.

If setting \code{usePlotly = FALSE}, the output of this function can be
further modified using standard ggplot2 syntax. If setting
\code{usePlotly = TRUE} an interactive plotly object will be produced.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# Draw the default plot for a single file
plotDupLevels(fdl[[1]])

plotDupLevels(fdl)

}
