% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFastqcPCA.R
\docType{methods}
\name{plotFastqcPCA}
\alias{plotFastqcPCA}
\alias{plotFastqcPCA,ANY-method}
\alias{plotFastqcPCA,character-method}
\alias{plotFastqcPCA,FastqcDataList-method}
\title{Draw a PCA plot for Fast QC modules}
\usage{
plotFastqcPCA(x, module, usePlotly = FALSE, labels, cluster = FALSE,
  clusterType = "colour", groups = NULL, ...)

\S4method{plotFastqcPCA}{ANY}(x, module, usePlotly = FALSE, labels,
  cluster = FALSE, clusterType = "colour", groups = NULL, ...)

\S4method{plotFastqcPCA}{character}(x, module, usePlotly = FALSE, labels,
  cluster = FALSE, clusterType = "colour", groups = NULL, ...)

\S4method{plotFastqcPCA}{FastqcDataList}(x, module, usePlotly = FALSE,
  labels, cluster = FALSE, clusterType = "colour", groups = NULL,
  ...)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{module}{\code{character} vector containing
the desired FastQC module (eg. c("Per_base_sequence_quality",
"Per_base_sequence_content"))}

\item{usePlotly}{\code{logical}. Output as ggplot2 (default) or plotly
object.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default}

\item{cluster}{\code{logical} default \code{FALSE}. If \code{groups} argument
is not set fastqc data will be clustered using hierarchical clustering.}

\item{clusterType}{One of "color/colour" or "hulls". Default is "colours"
and will colour points based on cluster/group, "hulls" will draw a polygon
around each cluster.}

\item{groups}{Optional data.frame (or tibble) with columns \code{Filename}
and \code{Group}. Values in the Filename column should correspond to the
values returned by fqName(x). If not supplied and \code{cluster = TRUE},
clusters will be automatically generated using HCPC from FactoMiner}

\item{...}{Used to pass additional attributes to theme() and between methods}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Draw a PCA plot for Fast QC modules across multiple samples
}
\details{
This carries out PCA on all or a subset of FastQC modules and plots the
output using either ggplot or plotly. Clustering of the PCA can be carried
out using a hierarchical clustering approach. Current modules for PCA are
Per_base_sequence_quality, Per_sequence_quality_scores,
Per_sequence_GC_content, Per_base_sequence_content, and
Sequence_Length_Distribution.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)
plotFastqcPCA(fdl, module = "Per_sequence_quality_scores", cluster = TRUE)


}
