% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGcContent.R
\docType{methods}
\name{plotGcContent}
\alias{plotGcContent}
\alias{plotGcContent,character-method}
\alias{plotGcContent,character}
\alias{plotGcContent,FastqcFile-method}
\alias{plotGcContent,FastqcFile}
\alias{plotGcContent,FastqcFileList-method}
\alias{plotGcContent,FastqcFileList}
\alias{plotGcContent,FastqcData-method}
\alias{plotGcContent,FastqcData}
\alias{plotGcContent,FastqcDataList-method}
\alias{plotGcContent,FastqcDataList}
\title{Plot the Per Sequence GC Content}
\usage{
plotGcContent(x, usePlotly = FALSE, labels, theoreticalGC = TRUE,
  theoreticalType = "Genome", species = "Hsapiens", GCobject, Fastafile,
  n = 1e+06, bp = 100, ...)

\S4method{plotGcContent}{character}(x, usePlotly = FALSE, labels,
  theoreticalGC = TRUE, theoreticalType = "Genome", species = "Hsapiens",
  GCobject, Fastafile, n = 1e+06, bp = 100, ...)

\S4method{plotGcContent}{FastqcFile}(x, usePlotly = FALSE, labels,
  theoreticalGC = TRUE, theoreticalType = "Genome", species = "Hsapiens",
  GCobject, Fastafile, n = 1e+06, bp = 100, ...)

\S4method{plotGcContent}{FastqcFileList}(x, usePlotly = FALSE, labels,
  theoreticalGC = TRUE, theoreticalType = "Genome", species = "Hsapiens",
  GCobject, Fastafile, n = 1e+06, bp = 100, ...)

\S4method{plotGcContent}{FastqcData}(x, usePlotly = FALSE, labels,
  theoreticalGC = TRUE, theoreticalType = "Genome", species = "Hsapiens",
  GCobject, Fastafile, n = 1e+06, bp = 100, counts = FALSE,
  lineCols = c("red", "blue"), ...)

\S4method{plotGcContent}{FastqcDataList}(x, usePlotly = FALSE, labels,
  theoreticalGC = TRUE, theoreticalType = "Genome", species = "Hsapiens",
  GCobject, Fastafile, n = 1e+06, bp = 100, plotType = c("heatmap",
  "line"), pwfCols, cluster = FALSE, dendrogram = TRUE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{theoreticalGC}{\code{logical} default is \code{FALSE} to give the true GC content%, set to \code{TRUE} to normalize
values of GC_Content by the theoretical values using \code{\link{gcTheoretical}}. \code{species} must be specified.}

\item{theoreticalType}{\code{character} Select type of data to normalize GC content% agianst accepts either "Genome" or
"Transcriptome". Default is "Genome"}

\item{species}{\code{character} if \code{gcTheory} is \code{TRUE} its must be accompanied by a species
Species currently supported can be obtained using \code{mData(gcTheoretical)}}

\item{GCobject}{an object of class GCTheoretical.
Defaults to the gcTheoretical object supplied witht= the package}

\item{Fastafile}{a fasta file contains DNA sequences to generate theoretical GC content}

\item{n}{number of simulated reads to generate theoretical GC content from \code{Fastafile}}

\item{bp}{simulated read length to generate theoretical GC content from \code{Fastafile}}

\item{...}{Used to pass various potting parameters to theme.}

\item{counts}{\code{logical}. Plot the counts from each file if \code{counts = TRUE}.
If \code{counts = FALSE} the frequencies will be plotted}

\item{lineCols}{Colors for observed and theoretical GC lines in single plots}

\item{plotType}{Takes values "line" or "heatmap"}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for pass, warning, and fail
values in plot}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}
}
\value{
A ggplot2 or plotly object
}
\description{
Plot the Per Sequence GC Content for a set of FASTQC files
}
\details{
Makes plots for GC_Content.
When applied to a single FastqcFile or FastqcData object a simple line plot will be drawn,
with Theoretical GC content overlaid if desired

For a FastqcFileList, or FastqcDataList either a line plot or heatmap can be drawn
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot for a FastqcDataList
plotGcContent(fdl)

# Plot a single FastqcData object
plotGcContent(fdl[[1]])

# Plot GC content with theoretical GC content generated from a given fasta file
plotGcContent(fdl, Fastafile = system.file("extdata","Athaliana.TAIR10.tRNA.fasta",
package="ngsReports"))
}
