% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGcContent.R
\name{plotGcContent}
\alias{plotGcContent}
\title{Plot the Per Sequence GC Content}
\usage{
plotGcContent(x, subset, counts = FALSE, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*", usePlotly = FALSE)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{counts}{\code{logical}. Plot the counts from each file if \code{counts = TRUE}.
If \code{counts = FALSE} the frequencies will be plotted}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileName}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileName.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}
}
\value{
A ggplot2 object
}
\description{
Plot the Per Sequence GC Content for a set of FASTQC files
}
\details{
Draws a plot of all GC Content overlaid across a single plot.
The mean from the observed data at each value is overlaid as a dotted black line.
Addition of a theoretical distribution has not yet been implemented.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# The default plot
plotGcContent(fdl)

# Plot the R1 files using counts
r1 <- grepl("R1", fileName(fdl))
plotGcContent(fdl, subset = r1 , counts = TRUE)



}
