% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKmerHeatmap.R
\name{plotKmerHeatmap}
\alias{plotKmerHeatmap}
\title{Plot Overrepresented Kmers As a Heatmap}
\usage{
plotKmerHeatmap(x, subset, nKmers = 12, low = rgb(0.2, 0, 0.2),
  mid = rgb(1, 0, 0), high = rgb(1, 1, 0), trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{nKmers}{\code{numeric}.
The number of Kmers to show.
Kmers are sorted by \code{max(Obs/Exp_Max)} acros all files to select these}

\item{low}{colour used as the low colour in the heatmap}

\item{mid}{colour used as the mid colour in the heatmap}

\item{high}{colour used as thehigh colour in the heatmap}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot Overrepresented Kmers across an entire library
}
