% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKmerHeatmap.R
\name{plotKmerHeatmap}
\alias{plotKmerHeatmap}
\title{Plot Overrepresented Kmers As a Heatmap}
\usage{
plotKmerHeatmap(x, subset, nKmers = 12, method = "overall", pwfCols,
  naCol = "grey80", flip = TRUE, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{nKmers}{\code{numeric}. The number of Kmers to show.}

\item{method}{Can only take the values \code{"overall"} or \code{"individual"}.
Determines whether the top nKmers are selected by the overall ranking (based on Obs/Exp_Max),
or whether the top nKmers are selected from each individual file.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{naCol}{colour used for missing values}

\item{flip}{\code{logical}. Enable a call to \code{coord_flip} to determine the best direction}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileName}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileName.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot Overrepresented Kmers across an entire library
}
\details{
Plots the over-represented Kmers as a heatmap.
Colours will correspond to the p-values.
In the presence of only zero-valued p-values the heatmap will be filled in a single colour,
with missing values defined simply as p < 0.

If non-zero p-values are present, colours will denote -log10 transformed values.
In this case zero values will be transformed using -log10(q) + 1,
where q is the minimum of the non-zero p-values.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Not a great example
plotKmerHeatmap(fdl)

# Find the R1 files with CC barcodes & just plot these
ccR1 <- grepl("CC.+R1", fileName(fdl))
plotKmerHeatmap(fdl, subset = ccR1, method = "individual", nKmers = 3)


}
