% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKmers.R
\docType{methods}
\name{plotKmers}
\alias{plotKmers}
\alias{plotKmers,ANY-method}
\alias{plotKmers,character-method}
\alias{plotKmers,FastqcData-method}
\alias{plotKmers,FastqcDataList-method}
\title{Plot Overrepresented Kmers}
\usage{
plotKmers(x, usePlotly = FALSE, labels, ...)

\S4method{plotKmers}{ANY}(x, usePlotly = FALSE, labels, ...)

\S4method{plotKmers}{character}(x, usePlotly = FALSE, labels, ...)

\S4method{plotKmers}{FastqcData}(
  x,
  usePlotly = FALSE,
  labels,
  n = 6,
  ...,
  lineWidth = 0.5,
  pal = c("red", "blue", "green", "black", "magenta", "yellow")
)

\S4method{plotKmers}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  cluster = FALSE,
  dendrogram = FALSE,
  pwfCols,
  heatCol = inferno(50),
  ...
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}

\item{n}{\code{numeric}. The number of Kmers to show.}

\item{lineWidth}{Passed to \code{geom_line(size = lineWidth)}}

\item{pal}{The colour palette. If the vector supplied is less than n,
\code{grDevices::colorRampPalette()} will be used}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for
pass, warning, and fail values in the plot}

\item{heatCol}{Colour palette used for the heatmap. Default is \code{inferno}
from the package \code{viridris}}
}
\value{
A standard ggplot2 object or an interactive plotly object
}
\description{
Plot Overrepresented Kmers
}
\details{
As the Kmer Content module present in FastQC reports is relatively
uninformative, and omitted by default in later versions of FastQC, these
are rudimentary plots.

Plots for \code{FastqcData} objects replicate those contained in a FastQC
report, whilst the heatmap generated from \code{FastqcDataList} objects
simply show the location and abundance of over-represented Kmers.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)
plotKmers(fdl[[1]])

}
