% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKmers.R
\docType{methods}
\name{plotKmers}
\alias{plotKmers}
\alias{plotKmers,character-method}
\alias{plotKmers,character}
\alias{plotKmers,FastqcFile-method}
\alias{plotKmers,FastqcFile}
\alias{plotKmers,FastqcFileList-method}
\alias{plotKmers,FastqcFileList}
\alias{plotKmers,FastqcData-method}
\alias{plotKmers,FastqcData}
\alias{plotKmers,FastqcDataList-method}
\alias{plotKmers,FastqcDataList}
\title{Plot Overrepresented Kmers}
\usage{
plotKmers(x, usePlotly = FALSE, ...)

\S4method{plotKmers}{character}(x, usePlotly = FALSE, ...)

\S4method{plotKmers}{FastqcFile}(x, usePlotly = FALSE, ...)

\S4method{plotKmers}{FastqcFileList}(x, usePlotly = FALSE, ...)

\S4method{plotKmers}{FastqcData}(x, usePlotly = FALSE, n = 6, labels,
  ..., lineWidth = 0.5, pal = c("red", "blue", "green", "black",
  "magenta", "yellow"))

\S4method{plotKmers}{FastqcDataList}(x, usePlotly = FALSE, labels,
  cluster = FALSE, dendrogram = FALSE, pwfCols,
  heatCol = inferno(50), ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{...}{Used to pass various potting parameters to theme.
Can also be used to set size and colour for box outlines.}

\item{n}{\code{numeric}. The number of Kmers to show.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{lineWidth}{Passed to \code{geom_line(size = lineWidth)}}

\item{pal}{The colour palette.
If the vector supplied is less than n, \code{grDevices::colorRampPalette()} will be used}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for pass, warning, and fail
values in plot}

\item{heatCol}{Colour palette used for the heatmap. Default is \code{inferno} from the package \code{viridris}}
}
\value{
A standard ggplot2 object or an interactive plotly object
}
\description{
Plot Overrepresented Kmers
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)
plotKmers(fdl[[1]])

}
