% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKmers.R
\name{plotKmers}
\alias{plotKmers}
\title{Plot Overrepresented Kmers}
\usage{
plotKmers(x, subset, nc = 2, nKmers = 6, method = "overall",
  trimNames = TRUE, pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{nc}{\code{numeric}. The number of columns to create in the plot layout}

\item{nKmers}{\code{numeric}. The number of Kmers to show.}

\item{method}{Can only take the values \code{"overall"} or \code{"individual"}.
Determines whether the top nKmers are selected by the overall ranking (based on Obs/Exp_Max),
or whether the top nKmers are selected from each individual file.}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot Overrepresented Kmers across an entire library
}
\details{
This plots the top overrepresented Kmers for one or more FASTQC reports.
Sequences are coloured in order of first appearance from left to right for easy
identification of larger sequences.

Returned plots can be easily modified using the standard ggplot2 methods
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "CAGG", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "fastqcReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Not a great example
plotKmers(fdl)

# Try digging a bit deeper
ccR1 <- grepl("CC.+R1", fileNames(fdl))
plotKmers(fdl, subset = ccR1, nc = 1, method = "individual", nKmers = 4)

}
