% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNContent.R
\docType{methods}
\name{plotNContent}
\alias{plotNContent}
\alias{plotNContent,character-method}
\alias{plotNContent,character}
\alias{plotNContent,FastqcFile-method}
\alias{plotNContent,FastqcFile}
\alias{plotNContent,FastqcFileList-method}
\alias{plotNContent,FastqcFileList}
\alias{plotNContent,FastqcData-method}
\alias{plotNContent,FastqcData}
\alias{plotNContent,FastqcDataList-method}
\alias{plotNContent,FastqcDataList}
\title{Draw an N Content Plot}
\usage{
plotNContent(x, usePlotly = FALSE, labels, pwfCols, warn = 5,
  fail = 20, ...)

\S4method{plotNContent}{character}(x, usePlotly = FALSE, labels, pwfCols,
  warn = 5, fail = 20, ...)

\S4method{plotNContent}{FastqcFile}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 5, fail = 20, ...)

\S4method{plotNContent}{FastqcFileList}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 5, fail = 20, ...)

\S4method{plotNContent}{FastqcData}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 5, fail = 20, ..., lineCol = "red")

\S4method{plotNContent}{FastqcDataList}(x, usePlotly = FALSE, labels,
  pwfCols, warn = 5, fail = 20, cluster = FALSE,
  dendrogram = FALSE, ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList},
\code{FastqcData}, \code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. Output as ggplot2 (default) or plotly
object.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours
for PASS/WARN/FAIL}

\item{warn, fail}{The default values for warn and fail are 5 and 10
respectively (i.e. percentages)}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{lineCol}{Defaults to red}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Draw an N Content Plot across one or more FASTQC reports
}
\details{
This extracts the N_Content from the supplied object and generates a ggplot2
object, with a set of minimal defaults.
The output of this function can be further modified using the standard
ggplot2 methods.

When \code{x} is a single FastqcFile, or FastqcData object line plots will
always be drawn for all Ns.
Otherwise, users can select line plots or heatmaps.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# The default plot
plotNContent(fdl[[1]])


}
