% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNContent.R
\docType{methods}
\name{plotNContent}
\alias{plotNContent}
\alias{plotNContent,ANY-method}
\alias{plotNContent,FastqcData-method}
\alias{plotNContent,FastqcDataList-method}
\alias{plotNContent,FastpData-method}
\alias{plotNContent,FastpDataList-method}
\title{Draw an N Content Plot}
\usage{
plotNContent(x, usePlotly = FALSE, labels, ...)

\S4method{plotNContent}{ANY}(x, usePlotly = FALSE, labels, pwfCols, warn = 5, fail = 20, ...)

\S4method{plotNContent}{FastqcData}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 5,
  fail = 20,
  ...,
  lineCol = "red"
)

\S4method{plotNContent}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  warn = 5,
  fail = 20,
  cluster = FALSE,
  dendrogram = FALSE,
  heat_w = 8,
  ...
)

\S4method{plotNContent}{FastpData}(
  x,
  usePlotly = FALSE,
  labels,
  pattern = ".fastq.*|.fastp.*|.fq*",
  module = c("Before_filtering", "After_filtering"),
  reads = c("read1", "read2"),
  linetype = NULL,
  colour = NULL,
  lineCol = c("navyblue", "red3"),
  facetBy = module ~ fqName,
  ...
)

\S4method{plotNContent}{FastpDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pattern = ".fastq.*|.fastp.*|.fq*",
  module = c("Before_filtering", "After_filtering"),
  reads = c("read1", "read2"),
  fill = scale_fill_viridis_c(),
  ...
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{usePlotly}{\code{logical}. Output as ggplot2 (default) or plotly
object.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{pwfCols}{Object of class \code{\link[=PwfCols]{PwfCols()}} containing the colours
for PASS/WARN/FAIL}

\item{warn, fail}{The default values for warn and fail are 5 and 10
respectively (i.e. percentages)}

\item{lineCol}{Line colours}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{heat_w}{Relative width of any heatmap plot components}

\item{pattern}{Regex used to trim filenames}

\item{module}{Used for Fastp* structures to show results before or after
filtering}

\item{reads}{Show plots for read1, read2 or both.}

\item{colour, linetype}{Data arguments to decorate plots. Should be drawn
from columns in the formatted data, such as "module", "reads", "fqName"}

\item{facetBy}{Formula to facet the plot by. Passed to \code{facet_wrap()}}

\item{fill}{continuous scale for ggplot objects}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Draw an N Content Plot across one or more FastQC reports
}
\details{
This extracts the N_Content from the supplied object and generates a ggplot2
object, with a set of minimal defaults.
The output of this function can be further modified using the standard
ggplot2 methods.

When \code{x} is a single FastqcData object line plots will always be drawn
for all Ns.
Otherwise, users can select line plots or heatmaps.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# The default plot
plotNContent(fdl[[1]])

}
