% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverrepresentedHeatmap.R
\name{plotOverrepresentedHeatmap}
\alias{plotOverrepresentedHeatmap}
\title{Plot Overrepresented Sequnces As a Heatmap}
\usage{
plotOverrepresentedHeatmap(x, subset, nSeq = 20, type = ".+",
  method = "overall", low = rgb(0.2, 0, 0.2), high = rgb(1, 0, 0),
  naCol = "grey80", flip = TRUE, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{nSeq}{\code{numeric}.
The number of Sequences to show.
Sequences are sorted by \code{max(Percentage)} across all files during selection}

\item{type}{A regular expression used to subset based on the Possible_Source column}

\item{method}{Can only take the values \code{"overall"} or \code{"individual"}.
Determines whether the top Sequences are selected by the overall ranking (based on Percentage),
or whether the top nKmers are selected from each individual file.}

\item{low}{colour used as the low colour in the heatmap}

\item{high}{colour used as the high colour in the heatmap}

\item{naCol}{colour used for missing values}

\item{flip}{\code{logical}. Enable a call to \code{coord_flip} to determine the best direction}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileName}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileName.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Plot Overrepresented Sequences across an entire library
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Another example which isn't ideal
plotOverrepresentedHeatmap(fdl)

# Dig a bit deeper
r1 <- grepl("R1", fileName(fdl))
plotOverrepresentedHeatmap(fdl, subset = r1, flip = FALSE, nSeq = 10)

# Check the top 2 sequences with No Hit from each R1 file
plotOverrepresentedHeatmap(fdl, subset = r1, type = "No Hit", nSeq = 2, method = "individual")

}
