% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOverrepresentedHeatmapPlotly.R
\name{plotOverrepresentedHeatmapPlotly}
\alias{plotOverrepresentedHeatmapPlotly}
\title{Plot Overrepresented Sequnces As a Heatmap}
\usage{
plotOverrepresentedHeatmapPlotly(x, subset, pwfCols, nSeq = 5, type = ".+",
  method = "Individual", low = rgb(0.2, 0, 0.2), high = rgb(1, 0, 0),
  naCol = "grey80", pattern = "(.+)\\\\.(fastq|fq).*",
  clusterNames = FALSE, usePlotly = FALSE, trimNames = TRUE,
  dendrogram = FALSE)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{pwfCols}{Object of class \code{PwfCols} to give colours for pass, warning, and fail
values in plot}

\item{nSeq}{\code{numeric}.
The number of Sequences to show.
Sequences are sorted by \code{max(Percentage)} across all files during selection}

\item{type}{A regular expression used to subset based on the Possible_Source column}

\item{method}{Can only take the values \code{"overall"} or \code{"individual"}.
Determines whether the top Sequences are selected by the overall ranking (based on Percentage),
or whether the top nKmers are selected from each individual file.}

\item{low}{colour used as the low colour in the heatmap}

\item{high}{colour used as the high colour in the heatmap}

\item{naCol}{colour used for missing values}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileName.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{clusterNames}{\code{logical} Cluster heatmap, or not.}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileName}

\item{dendrogram}{\code{logical} redundant if \code{clusterNames} is \code{FALSE}
if both \code{clusterNames} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}
}
\description{
Plot Overrepresented Sequences across an entire library
}
