% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadTotals.R
\docType{methods}
\name{plotReadTotals}
\alias{plotReadTotals}
\alias{plotReadTotals,character-method}
\alias{plotReadTotals,character}
\alias{plotReadTotals,FastqcFileList-method}
\alias{plotReadTotals,FastqcFileList}
\alias{plotReadTotals,FastqcDataList-method}
\alias{plotReadTotals,FastqcDataList}
\title{Draw a barplot of read totals}
\usage{
plotReadTotals(x, usePlotly = FALSE, duplicated = TRUE, bars = "stacked",
  ...)

\S4method{plotReadTotals}{character}(x, usePlotly = FALSE,
  duplicated = TRUE, bars = "stacked", ...)

\S4method{plotReadTotals}{FastqcFileList}(x, usePlotly = FALSE,
  duplicated = TRUE, bars = "stacked", ...)

\S4method{plotReadTotals}{FastqcDataList}(x, usePlotly = FALSE,
  duplicated = TRUE, bars = "stacked", labels, millions, barCols,
  expand.x = c(0, 0), expand.y = c(0, 0), ...)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}

\item{duplicated}{\code{logical}. Include deduplicated read total estimates to plot charts}

\item{bars}{If \code{duplicated = TRUE}, show unique and deduplicated reads as "stacked" or "adjacent".}

\item{...}{Used to pass additional attributes to theme()}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{millions}{\code{logical}. Use Millions of reads as the scale for the y-axis.
Unless specified, will be set as TRUE automatically if the highest total is > 2e06.}

\item{barCols}{Colours for duplicated and unique reads.}

\item{expand.x, expand.y}{Vectors of length 2. Passed to `scale_*_continuous()`}
}
\value{
Returns a ggplot or plotly object
}
\description{
Draw a barplot of read totals
}
\details{
Draw a barplot of read totals using the standard ggplot2 syntax.
Read totals will be plotted in millions as this is the most common.
The raw data from \code{\link{readTotals}} can otherwise be used to manually create a plot.

However, this is based on the value shown on FASTQC reports at the top of DeDuplicatedTotals plot,
and is known to be inaccurate.
As it still gives a good guide as to sequence diversity it is included as the default.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Plot the Read Totals showing estimated duplicates
plotReadTotals(fdl)

# Plot the Read Totals without estimated duplicates
plotReadTotals(fdl, duplicated = FALSE)

}
