% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadTotals.R
\docType{methods}
\name{plotReadTotals}
\alias{plotReadTotals}
\alias{plotReadTotals,ANY-method}
\alias{plotReadTotals,FastqcDataList-method}
\alias{plotReadTotals,FastpDataList-method}
\title{Draw a barplot of read totals}
\usage{
plotReadTotals(
  x,
  usePlotly = FALSE,
  labels,
  pattern = ".(fast|fq|bam|sam).*",
  ...
)

\S4method{plotReadTotals}{ANY}(
  x,
  usePlotly = FALSE,
  labels,
  pattern = ".(fast|fq|bam|sam).*",
  ...
)

\S4method{plotReadTotals}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pattern = ".(fast|fq|bam|sam).*",
  duplicated = TRUE,
  bars = c("stacked", "adjacent"),
  barCols = c("red", "blue"),
  expand.x = c(0, 0.02),
  ...
)

\S4method{plotReadTotals}{FastpDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pattern = ".(fast|fq|bam|sam).*",
  scalePaired = TRUE,
  scaleY = 1e+06,
  fillScale = NULL,
  labMin = 0.05,
  labVjust = 0.5,
  labFill = "white",
  plotTheme = theme(),
  expand.y = c(0, 0.05),
  ...
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.}

\item{pattern}{Regex used to trim the end of filenames}

\item{...}{Used to pass additional attributes to theme()}

\item{duplicated}{\code{logical}. Include deduplicated read total estimates
to plot charts}

\item{bars}{If \code{duplicated = TRUE}, show unique and deduplicated reads
as "stacked" or "adjacent".}

\item{barCols}{Colours for duplicated and unique reads.}

\item{expand.x, expand.y}{Passed to \code{\link[ggplot2:expansion]{ggplot2::expansion}} inside the
respective axis scale}

\item{scalePaired}{Scale read totals by 0.5 when paired}

\item{scaleY}{Scale read totals by this value. The default shows the y-axis
in millions}

\item{fillScale}{ScaleDiscrete function to be applied to the plot}

\item{labMin}{Only show labels for filtering categories higher than this
values as a proportion of reads}

\item{labVjust}{Used to place labels within each bar.}

\item{labFill}{Passed to \link[ggplot2]{geom_label}}

\item{plotTheme}{\link[ggplot2]{theme} to be added to the plot}
}
\value{
Returns a ggplot or plotly object
}
\description{
Draw a barplot of read totals
}
\details{
Draw a barplot of read totals using the standard ggplot2 syntax.
The raw data from \code{\link[=readTotals]{readTotals()}} can otherwise be used to manually
create a plot.

Duplication levels are based on the value shown on FASTQC reports at the
top of the DeDuplicatedTotals plot, which is known to be inaccurate.
As it still gives a good guide as to sequence diversity it is included as
the default. This can be turned off by setting \code{duplicated = FALSE}.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# Plot the Read Totals showing estimated duplicates
plotReadTotals(fdl)

# Plot the Read Totals without estimated duplicates
plotReadTotals(fdl, duplicated = FALSE)

}
