% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadTotals.R
\docType{methods}
\name{plotReadTotals}
\alias{plotReadTotals}
\alias{plotReadTotals,ANY-method}
\alias{plotReadTotals,character-method}
\alias{plotReadTotals,FastqcDataList-method}
\title{Draw a barplot of read totals}
\usage{
plotReadTotals(
  x,
  usePlotly = FALSE,
  labels,
  duplicated = TRUE,
  bars = c("stacked", "adjacent"),
  barCols = c("red", "blue"),
  expand.x = expansion(mult = c(0, 0.02)),
  ...
)

\S4method{plotReadTotals}{ANY}(
  x,
  usePlotly = FALSE,
  labels,
  duplicated = TRUE,
  bars = c("stacked", "adjacent"),
  barCols = c("red", "blue"),
  expand.x = expansion(mult = c(0, 0.02)),
  ...
)

\S4method{plotReadTotals}{character}(
  x,
  usePlotly = FALSE,
  labels,
  duplicated = TRUE,
  bars = c("stacked", "adjacent"),
  barCols = c("red", "blue"),
  expand.x = expansion(mult = c(0, 0.02)),
  ...
)

\S4method{plotReadTotals}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  duplicated = TRUE,
  bars = c("stacked", "adjacent"),
  barCols = c("red", "blue"),
  expand.x = expansion(mult = c(0, 0.02)),
  ...
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using
ggplot. If \code{TRUE} plot will be rendered with plotly}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{duplicated}{\code{logical}. Include deduplicated read total estimates
to plot charts}

\item{bars}{If \code{duplicated = TRUE}, show unique and deduplicated reads
as "stacked" or "adjacent".}

\item{barCols}{Colours for duplicated and unique reads.}

\item{expand.x}{Output from \code{expansion()} controlling x-axis
expansion. Alternatively can be a numeric vector of length 4}

\item{...}{Used to pass additional attributes to theme()}
}
\value{
Returns a ggplot or plotly object
}
\description{
Draw a barplot of read totals
}
\details{
Draw a barplot of read totals using the standard ggplot2 syntax.
The raw data from \code{\link{readTotals}} can otherwise be used to manually
create a plot.

Duplication levels are based on the value shown on FASTQC reports at the
top of the DeDuplicatedTotals plot, which is known to be inaccurate.
As it still gives a good guide as to sequence diversity it is included as
the default. This can be turned off by setting \code{duplicated = FALSE}.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# Plot the Read Totals showing estimated duplicates
plotReadTotals(fdl)

# Plot the Read Totals without estimated duplicates
plotReadTotals(fdl, duplicated = FALSE)

}
