% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadTotals.R
\name{plotReadTotals}
\alias{plotReadTotals}
\title{Draw a barplot of read totals}
\usage{
plotReadTotals(x, subset, millions, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*", usePlotly = FALSE)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{millions}{\code{logical}. Use Millions of reads as the scale for the y-axis.
Unless specified, will be set as TRUE automatically if the highest total is > 2e06.}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileName}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileName.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{usePlotly}{\code{logical} Default \code{FALSE} will render using ggplot.
If \code{TRUE} plot will be rendered with plotly}
}
\value{
Returns a ggplot object.
}
\description{
Draw a barplot of read totals
}
\details{
Draw a barplot of read totals using the standard ggplot2 syntax.
Read totals will be plotted in millions as this is the most common.
The raw data from \code{\link{readTotals}} can otherwise be used to manually create a plot.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Plot the Read Totals
plotReadTotals(fdl)

# Change the scale so it is not in millions
# Also subset the reads to just the R1 files
r1 <- grepl("R1", fileName(fdl))
plotReadTotals(fdl, subset = r1, millions = FALSE)

}
