% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReadTotals.R
\name{plotReadTotals}
\alias{plotReadTotals}
\title{Draw a barplot of read totals}
\usage{
plotReadTotals(x, subset, trimNames = TRUE,
  pattern = "(.+)\\\\.(fastq|fq).*", th)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful to only return totals from R1 files, or any other subset}

\item{trimNames}{\code{logical}. Capture the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}

\item{th}{A \code{ggplot2 theme} object. Defaults to \code{theme_bw()}}
}
\description{
Draw a barplot of read totals
}
\details{
Draw a barplot of read totals using the standard ggplot2 syntax
}
