% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeqContent.R
\docType{methods}
\name{plotSeqContent}
\alias{plotSeqContent}
\alias{plotSeqContent,ANY-method}
\alias{plotSeqContent,character-method}
\alias{plotSeqContent,FastqcData-method}
\alias{plotSeqContent,FastqcDataList-method}
\title{Plot the per base content as a heatmap}
\usage{
plotSeqContent(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{ANY}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{character}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{FastqcData}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqContent}{FastqcDataList}(
  x,
  usePlotly = FALSE,
  labels,
  pwfCols,
  plotType = c("heatmap", "line", "residuals"),
  cluster = FALSE,
  dendrogram = FALSE,
  ...,
  nc = 2
)
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{usePlotly}{\code{logical}. Generate an interactive plot using plotly}

\item{labels}{An optional named vector of labels for the file names.
All file names must be present in the names of the vector.
File extensions are dropped by default.}

\item{...}{Used to pass additional attributes to theme() and between methods}

\item{pwfCols}{Object of class \code{\link{PwfCols}} to give colours for
pass, warning, and fail
values in plot}

\item{plotType}{\code{character}. Type of plot to generate. Must be "line",
"heatmap" or "residuals"}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE}
then the dendrogram will be displayed.}

\item{nc}{Specify the number of columns if plotting a FastqcDataList as line
plots. Passed to \code{ggplot2::facet_wrap}.}
}
\value{
A ggplot2 object or an interactive plotly object
}
\description{
Plot the Per Base content for a set of FASTQC files.
}
\details{
Per base sequence content (%A, %T, %G, %C), is shown as four overlaid
heatmap colours when plotting from multiple reports. The individual line
plots are able to be generated by setting \code{plotType = "line"}, and the
layout is determined by \code{facet_wrap} from ggplot2.

Individual line plots are also generated when plotting from a single
\code{FastqcData} object.

If setting \code{usePlotly = TRUE} for a large number of reports, the plot
can be slow to render.
An alternative may be to produce a plot of residuals for each base, produced
by taking the position-specific mean for each base.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# The default plot
plotSeqContent(fdl)

}
