% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeqLengthDistn.R
\docType{methods}
\name{plotSeqLengthDistn}
\alias{plotSeqLengthDistn}
\alias{plotSeqLengthDistn,ANY-method}
\alias{plotSeqLengthDistn,character-method}
\alias{plotSeqLengthDistn,FastqcData-method}
\alias{plotSeqLengthDistn,FastqcDataList-method}
\title{Plot the Sequence Length Distribution}
\usage{
plotSeqLengthDistn(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqLengthDistn}{ANY}(x, usePlotly = FALSE, labels, ...)

\S4method{plotSeqLengthDistn}{character}(x, usePlotly = FALSE, labels,
  ...)

\S4method{plotSeqLengthDistn}{FastqcData}(x, usePlotly = FALSE, labels,
  plotType = c("line", "cumulative"), ..., expand.x = expand_scale(0,
  0.2))

\S4method{plotSeqLengthDistn}{FastqcDataList}(x, usePlotly = FALSE,
  labels, counts = FALSE, plotType = c("heatmap", "line",
  "cumulative"), cluster = FALSE, dendrogram = FALSE, ...,
  expand.x = expand_scale(0, 0.2), heatCol = inferno(50))
}
\arguments{
\item{x}{Can be a \code{FastqcData}, \code{FastqcDataList} or file paths}

\item{usePlotly}{\code{logical}. Output as ggplot2 or plotly object.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{...}{Used to pass additional attributes to theme()}

\item{plotType}{\code{character}. Can only take the values
\code{plotType = "heatmap"} \code{plotType = "line"} or
\code{plotType = "cumulative"}}

\item{expand.x}{Output from \code{expand_scale()} or numeric vector of
length 4. Passed to \code{scale_x_discrete}}

\item{counts}{\code{logical} Should distributions be shown as counts or
frequencies (percentages)}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} and
\code{usePlotly} are \code{FALSE}. If both \code{cluster} and
\code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{heatCol}{The colour scheme for the heatmap}
}
\value{
A standard ggplot2 object, or an interactive plotly object
}
\description{
Plot the Sequence Length Distribution across one or more FASTQC
reports
}
\details{
This extracts the Sequence Length Distribution from the supplied object and
generates a ggplot2 object, with a set of minimal defaults.
The output of this function can be further modified using the standard
ggplot2 methods.

A cumulative plot can also be generated to provide guidance for minimum
read length in some NGS workflows, by setting \code{plotType = "cumulative"}.
If all libraries have reads of identical lengths, these plots may be less
informative.

An alternative interactive plot is available by setting the argument
\code{usePlotly = TRUE}.
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fl <- list.files(packageDir, pattern = "fastqc.zip", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- FastqcDataList(fl)

# Plot as a frequency plot using lines
plotSeqLengthDistn(fdl)

# Or plot the cumulative value
plotSeqLengthDistn(fdl, plotType = "cumulative")

}
