% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummary.R
\docType{methods}
\name{plotSummary}
\alias{plotSummary}
\alias{plotSummary,character-method}
\alias{plotSummary,character}
\alias{plotSummary,FastqcFileList-method}
\alias{plotSummary,FastqcFileList}
\alias{plotSummary,FastqcDataList-method}
\alias{plotSummary,FastqcDataList}
\title{Plot the PASS/WARN/FAIL information}
\usage{
plotSummary(x, usePlotly = FALSE, ...)

\S4method{plotSummary}{character}(x, usePlotly = FALSE, ...)

\S4method{plotSummary}{FastqcFileList}(x, usePlotly = FALSE, ...)

\S4method{plotSummary}{FastqcDataList}(x, usePlotly = FALSE, labels, pwfCols,
  cluster = FALSE, dendrogram = TRUE, ..., gridlineWidth = 0.2,
  gridlineCol = "grey20")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{usePlotly}{\code{logical}. Generate an interactive plot using plotly}

\item{...}{Used to pass various potting parameters to theme.}

\item{labels}{An optional named vector of labels for the file names.
All filenames must be present in the names.
File extensions are dropped by default.}

\item{pwfCols}{Object of class \code{\link{PwfCols}} containing the colours for PASS/WARN/FAIL}

\item{cluster}{\code{logical} default \code{FALSE}. If set to \code{TRUE},
fastqc data will be clustered using hierarchical clustering}

\item{dendrogram}{\code{logical} redundant if \code{cluster} is \code{FALSE}
if both \code{cluster} and \code{dendrogram} are specified as \code{TRUE} then the dendrogram
will be displayed.}

\item{gridlineWidth, gridlineCol}{Passed to geom_hline and geom_vline to determine
width and colour of gridlines}
}
\value{
A ggplot2 object (\code{usePlotly = FALSE})
or an interactive plotly object (\code{usePlotly = TRUE})
}
\description{
Extract the PASS/WARN/FAIL summaries and plot them
}
\details{
This uses the standard ggplot2 syntax to create a three colour plot.
The output of this function can be further modified using the standard ggplot2 methods.
}
\examples{

# Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Check the overall PASS/WARN/FAIL status
plotSummary(fdl)

# Interactive plot
plotSummary(fdl, usePlotly = TRUE)

}
