% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummary.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot the PASS/WARN/FAIL information}
\usage{
plotSummary(x, subset, col = c(FAIL = "red", WARN = "yellow", PASS = "green"),
  trimNames = TRUE, pattern = "(.+)\\\\.(fastq|fq).*")
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful for only return totals from R1 files, or any other subset}

\item{col}{\code{character vector} of colours}

\item{trimNames}{\code{logical}. Remove the text specified in \code{pattern} from fileNames}

\item{pattern}{\code{character}.
Contains a regular expression which will be captured from fileNames.
The default will capture all text preceding .fastq/fastq.gz/fq/fq.gz}
}
\value{
A standard ggplot2 object
}
\description{
Extract the PASS/WARN/FAIL summaries and plot them
}
\details{
This uses the standard ggplot2 syntax to create a three colour plot.
The output of this function can be further modified using the standard ggplot2 methods.
}
