% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTotals.R
\name{readTotals}
\alias{readTotals}
\title{Get the read totals}
\usage{
readTotals(x)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}
}
\value{
A \code{data_frame} with the columns \code{Filename} and \code{Total_Sequences}
}
\description{
Get the read totals from one or more FASTQC reports
}
