% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTotals.R
\name{readTotals}
\alias{readTotals}
\title{Get the read totals}
\usage{
readTotals(x, subset, trimNames = TRUE)
}
\arguments{
\item{x}{Can be a \code{FastqcFile}, \code{FastqcFileList}, \code{FastqcData},
\code{FastqcDataList} or path}

\item{subset}{\code{logical}. Return the values for a subset of files.
May be useful for only return totals from R1 files, or any other subset}

\item{trimName}{\code{logical}. Remove the suffix from fileNames}
}
\value{
A \code{data_frame} with two columns only
}
\description{
Get the read totals from one or more FASTQC reports
}
