% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastQC.R
\docType{methods}
\name{runFastQC}
\alias{runFastQC}
\alias{runFastQC,character-method}
\alias{runFastQC,FastqFile-method}
\alias{runFastQC,FastqFileList-method}
\alias{runFastQC,BamFile-method}
\alias{runFastQC,BamFileList-method}
\title{A wrapper for the bash shell command fastqc.}
\usage{
runFastQC(object, outPath, exec = "/usr/local/bin/fastqc",
  threads = 1L, casava = FALSE, nofilter = FALSE, extract = FALSE,
  nogroup = FALSE, min_length = 1, contaminants = c(),
  adapters = c(), kmers = 7)

\S4method{runFastQC}{character}(object, outPath,
  exec = "/usr/local/bin/fastqc", threads = 1L, casava = FALSE,
  nofilter = FALSE, extract = FALSE, nogroup = FALSE,
  min_length = 1, contaminants = c(), adapters = c(), kmers = 7)

\S4method{runFastQC}{FastqFile}(object, outPath,
  exec = "/usr/local/bin/fastqc", threads = 1L, casava = FALSE,
  nofilter = FALSE, extract = FALSE, nogroup = FALSE,
  min_length = 1, contaminants = c(), adapters = c(), kmers = 7)

\S4method{runFastQC}{FastqFileList}(object, outPath,
  exec = "/usr/local/bin/fastqc", threads = 1L, casava = FALSE,
  nofilter = FALSE, extract = FALSE, nogroup = FALSE,
  min_length = 1, contaminants = c(), adapters = c(), kmers = 7)

\S4method{runFastQC}{BamFile}(object, outPath,
  exec = "/usr/local/bin/fastqc", threads = 1L, casava = FALSE,
  nofilter = FALSE, extract = FALSE, nogroup = FALSE,
  min_length = 1, contaminants = c(), adapters = c(), kmers = 7)

\S4method{runFastQC}{BamFileList}(object, outPath,
  exec = "/usr/local/bin/fastqc", threads = 1L, casava = FALSE,
  nofilter = FALSE, extract = FALSE, nogroup = FALSE,
  min_length = 1, contaminants = c(), adapters = c(), kmers = 7)
}
\arguments{
\item{object}{A \code{\link{FastqFileList}}}

\item{outPath}{The path to write the FastQC reports. Must exist as (for safety)
it will not be created when calling this function}

\item{exec}{The location of the fastqc executable.}

\item{threads}{The number of threads to run in parallel}

\item{casava}{logical. Sets the \code{--casava} flag}

\item{nofilter}{logical. Sets the \code{--nofilter} flag}

\item{extract}{logical. Extract the zip files on completion of the report}

\item{nogroup}{logical. Sets the grouping of bases for reads longer than 50bp}

\item{min_length}{integer. Sets an artificial lower limit on the length of the 
sequence to be shown in the report.}

\item{contaminants}{Path to an alternate file with contaminants.
The structure of the file will not be checked.
Refer to the \code{fastqc} help page for more details}

\item{adapters}{Path toa file listing adapters to search for.
The structure of the file will not be checked.
Refer to the \code{fastqc} help page for more details}

\item{kmers}{An integer between 2 and 10}
}
\value{
An object of class \code{FastqcFileList}
}
\description{
Takes a FastqFileList
}
\details{
This is a simple wrapper function for controlling & running 
\code{fastqc} from within R.
This can be very useful for controlling & documenting an entire pipeline from 
within knitr to produce a simple report

Only the common functinoality of FastQC is implemented,
for more fine detail pease call FastQC directly.
}
\examples{
\dontrun{
library(ShortRead)
sp <- SolexaPath(system.file('extdata', package='ShortRead'))
fl <- file.path(analysisPath(sp), "s_1_sequence.txt")
f <- FastqFile(fl)
# This requires a working installation of FastQC
fqcFile <- runFastQC(f, outPath = tempdir())
}

}
\seealso{
\code{\link{FastqFileList}}
}
\author{
Steve Pederson <stephen.pederson@adelaide.edu.au>
}
