% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseOverrep.R
\name{summariseOverrep}
\alias{summariseOverrep}
\alias{summariseOverrep,FastpData-method}
\alias{summariseOverrep,FastpDataList-method}
\title{Summarise Overrepresented Sequences}
\usage{
summariseOverrep(x, ...)

\S4method{summariseOverrep}{FastpData}(x, step = c("Before", "After"), min_count = 0, ...)

\S4method{summariseOverrep}{FastpDataList}(
  x,
  min_count = 0,
  step = c("Before", "After"),
  vals = c("count", "rate"),
  fn = c("mean", "sum", "max"),
  ...
)
}
\arguments{
\item{x}{An object of a suitable class}

\item{...}{Not used}

\item{step}{Can be 'Before', 'After' or both to obtain data from the
Before_filtering or After_filtering modules}

\item{min_count}{Filter sequences with counts less than this value, both
before and after filtering}

\item{vals}{Values to use for creating summaries across multiple files}

\item{fn}{Functions to use when summarising values across multiple files}
}
\value{
A tibble
}
\description{
Summarise the Overrepresented sequences found in one or more QC files
}
\details{
This function prepares a useful summary of all over-represented sequences
as reported by either fastp or FastQC
}
\examples{
f <- system.file("extdata/fastp.json.gz", package = "ngsReports")
fp <- FastpData(f)
summariseOverrep(fp, min_count = 100)


}
