% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version.R
\docType{methods}
\name{Version,FastqcData-method}
\alias{Version,FastqcData-method}
\alias{Version}
\alias{Version,FastqcDataList-method}
\alias{Version}
\title{Get the FASTQC version}
\usage{
\S4method{Version}{FastqcData}(object)

\S4method{Version}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class \code{FastqcData} or \code{FastqcDataList}}
}
\value{
A character vector
}
\description{
Get the FASTQC version used to generate the intial files
}
