% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version.R
\docType{methods}
\name{Version,FastqcData-method}
\alias{Version,FastqcData-method}
\alias{Version}
\alias{Version,FastqcDataList-method}
\title{Get the FASTQC version}
\usage{
\S4method{Version}{FastqcData}(object)

\S4method{Version}{FastqcDataList}(object)
}
\arguments{
\item{object}{An object of class \code{FastqcData} or \code{FastqcDataList}}
}
\value{
A character vector (FastqcData), or data_frame (FastqcDataList)
}
\description{
Get the FASTQC version used to generate the intial files
}
\examples{

# Get the files included with the package
packageDir <- system.file("extdata", package = "ngsReports")
fileList <- list.files(packageDir, pattern = "fastqc", full.names = TRUE)

# Load the FASTQC data as a FastqcDataList object
fdl <- getFastqcData(fileList)

# Get the FASTQC version
Version(fdl)

}
