% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHtmlReport.R
\name{writeHtmlReport}
\alias{writeHtmlReport}
\title{Write an HTML Summary Report}
\usage{
writeHtmlReport(fastqcDir, template, species = "Hsapiens",
  dataType = "Transcriptome", nOver = 30, nKmer = 30, targetsDF,
  overwrite = FALSE, quiet = TRUE)
}
\arguments{
\item{fastqcDir}{A directory containing zipped, or extracted FastQC reports}

\item{template}{The template file which will be copied into \code{fastqcDir}}

\item{species}{Species/closely related species of sequenced samples}

\item{dataType}{Is the data "Transcriptomic" or "Genomic" in nature?}

\item{nOver}{The maximum number of Overrepresented Sequences to show}

\item{nKmer}{The maximum number of Kmers to show}

\item{targetsDF}{A data.frame with at least two columns named
\code{Filename}  and \code{Label}.
The filenames should match the original fastq files, and the labels should
be simply alternative labels for these files for convenience.}

\item{overwrite}{\code{logical}. Overwrite any previous copies of the
template file in the destination directory}

\item{quiet}{\code{logical}. Show or hide markdown output in the Console.}
}
\value{
Silently returns \code{TRUE} and will output a compiled HTML file from the
supplied Rmarkdown template file
}
\description{
Compiles an HTML report using a supplied template
}
\details{
This will take a user supplied template, or the file supplied with the
package and create an HTML summary of all standard FASTQC plots for all
files in the supplied directory.
}
